Struct aws_sdk_medialive::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for MediaLive.

This client allows ergonomic access to a MediaLive-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_input_device_transfer(&self) -> AcceptInputDeviceTransfer<C, M, R>

pub fn batch_delete(&self) -> BatchDelete<C, M, R>

pub fn batch_start(&self) -> BatchStart<C, M, R>

pub fn batch_stop(&self) -> BatchStop<C, M, R>

pub fn batch_update_schedule(&self) -> BatchUpdateSchedule<C, M, R>

pub fn cancel_input_device_transfer(&self) -> CancelInputDeviceTransfer<C, M, R>

pub fn create_channel(&self) -> CreateChannel<C, M, R>

pub fn create_input(&self) -> CreateInput<C, M, R>

pub fn create_input_security_group(&self) -> CreateInputSecurityGroup<C, M, R>

pub fn create_multiplex(&self) -> CreateMultiplex<C, M, R>

pub fn create_multiplex_program(&self) -> CreateMultiplexProgram<C, M, R>

pub fn create_partner_input(&self) -> CreatePartnerInput<C, M, R>

pub fn create_tags(&self) -> CreateTags<C, M, R>

pub fn delete_channel(&self) -> DeleteChannel<C, M, R>

pub fn delete_input(&self) -> DeleteInput<C, M, R>

pub fn delete_input_security_group(&self) -> DeleteInputSecurityGroup<C, M, R>

pub fn delete_multiplex(&self) -> DeleteMultiplex<C, M, R>

pub fn delete_multiplex_program(&self) -> DeleteMultiplexProgram<C, M, R>

pub fn delete_reservation(&self) -> DeleteReservation<C, M, R>

pub fn delete_schedule(&self) -> DeleteSchedule<C, M, R>

pub fn delete_tags(&self) -> DeleteTags<C, M, R>

pub fn describe_channel(&self) -> DescribeChannel<C, M, R>

pub fn describe_input(&self) -> DescribeInput<C, M, R>

pub fn describe_input_device(&self) -> DescribeInputDevice<C, M, R>

pub fn describe_input_device_thumbnail(
&self
) -> DescribeInputDeviceThumbnail<C, M, R>

pub fn describe_input_security_group(
&self
) -> DescribeInputSecurityGroup<C, M, R>

pub fn describe_multiplex(&self) -> DescribeMultiplex<C, M, R>

pub fn describe_multiplex_program(&self) -> DescribeMultiplexProgram<C, M, R>

pub fn describe_offering(&self) -> DescribeOffering<C, M, R>

pub fn describe_reservation(&self) -> DescribeReservation<C, M, R>

pub fn describe_schedule(&self) -> DescribeSchedule<C, M, R>

pub fn list_channels(&self) -> ListChannels<C, M, R>

pub fn list_input_devices(&self) -> ListInputDevices<C, M, R>

pub fn list_input_device_transfers(&self) -> ListInputDeviceTransfers<C, M, R>

pub fn list_inputs(&self) -> ListInputs<C, M, R>

pub fn list_input_security_groups(&self) -> ListInputSecurityGroups<C, M, R>

pub fn list_multiplexes(&self) -> ListMultiplexes<C, M, R>

pub fn list_multiplex_programs(&self) -> ListMultiplexPrograms<C, M, R>

pub fn list_offerings(&self) -> ListOfferings<C, M, R>

pub fn list_reservations(&self) -> ListReservations<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn purchase_offering(&self) -> PurchaseOffering<C, M, R>

pub fn reject_input_device_transfer(&self) -> RejectInputDeviceTransfer<C, M, R>

pub fn start_channel(&self) -> StartChannel<C, M, R>

pub fn start_multiplex(&self) -> StartMultiplex<C, M, R>

pub fn stop_channel(&self) -> StopChannel<C, M, R>

pub fn stop_multiplex(&self) -> StopMultiplex<C, M, R>

pub fn transfer_input_device(&self) -> TransferInputDevice<C, M, R>

pub fn update_channel(&self) -> UpdateChannel<C, M, R>

pub fn update_channel_class(&self) -> UpdateChannelClass<C, M, R>

pub fn update_input(&self) -> UpdateInput<C, M, R>

pub fn update_input_device(&self) -> UpdateInputDevice<C, M, R>

pub fn update_input_security_group(&self) -> UpdateInputSecurityGroup<C, M, R>

pub fn update_multiplex(&self) -> UpdateMultiplex<C, M, R>

pub fn update_multiplex_program(&self) -> UpdateMultiplexProgram<C, M, R>

pub fn update_reservation(&self) -> UpdateReservation<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self