Module aws_sdk_medialive::input

Modules

accept_input_device_transfer_input

See AcceptInputDeviceTransferInput

batch_delete_input

See BatchDeleteInput

batch_start_input

See BatchStartInput

batch_stop_input

See BatchStopInput

batch_update_schedule_input

See BatchUpdateScheduleInput

cancel_input_device_transfer_input

See CancelInputDeviceTransferInput

create_channel_input

See CreateChannelInput

create_input_input

See CreateInputInput

create_input_security_group_input

See CreateInputSecurityGroupInput

create_multiplex_input

See CreateMultiplexInput

create_multiplex_program_input

See CreateMultiplexProgramInput

create_partner_input_input

See CreatePartnerInputInput

create_tags_input

See CreateTagsInput

delete_channel_input

See DeleteChannelInput

delete_input_input

See DeleteInputInput

delete_input_security_group_input

See DeleteInputSecurityGroupInput

delete_multiplex_input

See DeleteMultiplexInput

delete_multiplex_program_input

See DeleteMultiplexProgramInput

delete_reservation_input

See DeleteReservationInput

delete_schedule_input

See DeleteScheduleInput

delete_tags_input

See DeleteTagsInput

describe_channel_input

See DescribeChannelInput

describe_input_device_input

See DescribeInputDeviceInput

describe_input_device_thumbnail_input

See DescribeInputDeviceThumbnailInput

describe_input_input

See DescribeInputInput

describe_input_security_group_input

See DescribeInputSecurityGroupInput

describe_multiplex_input

See DescribeMultiplexInput

describe_multiplex_program_input

See DescribeMultiplexProgramInput

describe_offering_input

See DescribeOfferingInput

describe_reservation_input

See DescribeReservationInput

describe_schedule_input

See DescribeScheduleInput

list_channels_input

See ListChannelsInput

list_input_device_transfers_input

See ListInputDeviceTransfersInput

list_input_devices_input

See ListInputDevicesInput

list_input_security_groups_input

See ListInputSecurityGroupsInput

list_inputs_input

See ListInputsInput

list_multiplex_programs_input

See ListMultiplexProgramsInput

list_multiplexes_input

See ListMultiplexesInput

list_offerings_input

See ListOfferingsInput

list_reservations_input

See ListReservationsInput

list_tags_for_resource_input

See ListTagsForResourceInput

purchase_offering_input

See PurchaseOfferingInput

reject_input_device_transfer_input

See RejectInputDeviceTransferInput

start_channel_input

See StartChannelInput

start_multiplex_input

See StartMultiplexInput

stop_channel_input

See StopChannelInput

stop_multiplex_input

See StopMultiplexInput

transfer_input_device_input

See TransferInputDeviceInput

update_channel_class_input

See UpdateChannelClassInput

update_channel_input

See UpdateChannelInput

update_input_device_input

See UpdateInputDeviceInput

update_input_input

See UpdateInputInput

update_input_security_group_input

See UpdateInputSecurityGroupInput

update_multiplex_input

See UpdateMultiplexInput

update_multiplex_program_input

See UpdateMultiplexProgramInput

update_reservation_input

See UpdateReservationInput

Structs

AcceptInputDeviceTransferInput

Placeholder documentation for AcceptInputDeviceTransferRequest

BatchDeleteInput

A request to delete resources

BatchStartInput

A request to start resources

BatchStopInput

A request to stop resources

BatchUpdateScheduleInput

List of actions to create and list of actions to delete.

CancelInputDeviceTransferInput

Placeholder documentation for CancelInputDeviceTransferRequest

CreateChannelInput

A request to create a channel

CreateInputInput

The name of the input

CreateInputSecurityGroupInput

The IPv4 CIDRs to whitelist for this Input Security Group

CreateMultiplexInput

A request to create a multiplex.

CreateMultiplexProgramInput

A request to create a program in a multiplex.

CreatePartnerInputInput

A request to create a partner input

CreateTagsInput

Placeholder documentation for CreateTagsRequest

DeleteChannelInput

Placeholder documentation for DeleteChannelRequest

DeleteInputInput

Placeholder documentation for DeleteInputRequest

DeleteInputSecurityGroupInput

Placeholder documentation for DeleteInputSecurityGroupRequest

DeleteMultiplexInput

Placeholder documentation for DeleteMultiplexRequest

DeleteMultiplexProgramInput

Placeholder documentation for DeleteMultiplexProgramRequest

DeleteReservationInput

Placeholder documentation for DeleteReservationRequest

DeleteScheduleInput

Placeholder documentation for DeleteScheduleRequest

DeleteTagsInput

Placeholder documentation for DeleteTagsRequest

DescribeChannelInput

Placeholder documentation for DescribeChannelRequest

DescribeInputDeviceInput

Placeholder documentation for DescribeInputDeviceRequest

DescribeInputDeviceThumbnailInput

Placeholder documentation for DescribeInputDeviceThumbnailRequest

DescribeInputInput

Placeholder documentation for DescribeInputRequest

DescribeInputSecurityGroupInput

Placeholder documentation for DescribeInputSecurityGroupRequest

DescribeMultiplexInput

Placeholder documentation for DescribeMultiplexRequest

DescribeMultiplexProgramInput

Placeholder documentation for DescribeMultiplexProgramRequest

DescribeOfferingInput

Placeholder documentation for DescribeOfferingRequest

DescribeReservationInput

Placeholder documentation for DescribeReservationRequest

DescribeScheduleInput

Placeholder documentation for DescribeScheduleRequest

ListChannelsInput

Placeholder documentation for ListChannelsRequest

ListInputDeviceTransfersInput

Placeholder documentation for ListInputDeviceTransfersRequest

ListInputDevicesInput

Placeholder documentation for ListInputDevicesRequest

ListInputSecurityGroupsInput

Placeholder documentation for ListInputSecurityGroupsRequest

ListInputsInput

Placeholder documentation for ListInputsRequest

ListMultiplexProgramsInput

Placeholder documentation for ListMultiplexProgramsRequest

ListMultiplexesInput

Placeholder documentation for ListMultiplexesRequest

ListOfferingsInput

Placeholder documentation for ListOfferingsRequest

ListReservationsInput

Placeholder documentation for ListReservationsRequest

ListTagsForResourceInput

Placeholder documentation for ListTagsForResourceRequest

PurchaseOfferingInput

Placeholder documentation for PurchaseOfferingRequest

RejectInputDeviceTransferInput

Placeholder documentation for RejectInputDeviceTransferRequest

StartChannelInput

Placeholder documentation for StartChannelRequest

StartMultiplexInput

Placeholder documentation for StartMultiplexRequest

StopChannelInput

Placeholder documentation for StopChannelRequest

StopMultiplexInput

Placeholder documentation for StopMultiplexRequest

TransferInputDeviceInput

A request to transfer an input device.

UpdateChannelClassInput

Channel class that the channel should be updated to.

UpdateChannelInput

A request to update a channel.

UpdateInputDeviceInput

A request to update an input device.

UpdateInputInput

A request to update an input.

UpdateInputSecurityGroupInput

The request to update some combination of the Input Security Group name and the IPv4 CIDRs the Input Security Group should allow.

UpdateMultiplexInput

A request to update a multiplex.

UpdateMultiplexProgramInput

A request to update a program in a multiplex.

UpdateReservationInput

Request to update a reservation