Module aws_sdk_medialive::model

Modules

aac_settings

See AacSettings

ac3_settings

See Ac3Settings

ancillary_source_settings

See AncillarySourceSettings

archive_cdn_settings

See ArchiveCdnSettings

archive_container_settings

See ArchiveContainerSettings

archive_group_settings

See ArchiveGroupSettings

archive_output_settings

See ArchiveOutputSettings

archive_s3_settings

See ArchiveS3Settings

arib_destination_settings

See AribDestinationSettings

arib_source_settings

See AribSourceSettings

audio_channel_mapping

See AudioChannelMapping

audio_codec_settings

See AudioCodecSettings

audio_description

See AudioDescription

audio_language_selection

See AudioLanguageSelection

audio_normalization_settings

See AudioNormalizationSettings

audio_only_hls_settings

See AudioOnlyHlsSettings

audio_pid_selection

See AudioPidSelection

audio_selector

See AudioSelector

audio_selector_settings

See AudioSelectorSettings

audio_silence_failover_settings

See AudioSilenceFailoverSettings

audio_track

See AudioTrack

audio_track_selection

See AudioTrackSelection

automatic_input_failover_settings

See AutomaticInputFailoverSettings

avail_blanking

See AvailBlanking

avail_configuration

See AvailConfiguration

avail_settings

See AvailSettings

batch_failed_result_model

See BatchFailedResultModel

batch_schedule_action_create_request

See BatchScheduleActionCreateRequest

batch_schedule_action_create_result

See BatchScheduleActionCreateResult

batch_schedule_action_delete_request

See BatchScheduleActionDeleteRequest

batch_schedule_action_delete_result

See BatchScheduleActionDeleteResult

batch_successful_result_model

See BatchSuccessfulResultModel

blackout_slate

See BlackoutSlate

burn_in_destination_settings

See BurnInDestinationSettings

caption_description

See CaptionDescription

caption_destination_settings

See CaptionDestinationSettings

caption_language_mapping

See CaptionLanguageMapping

caption_rectangle

See CaptionRectangle

caption_selector

See CaptionSelector

caption_selector_settings

See CaptionSelectorSettings

cdi_input_specification

See CdiInputSpecification

channel

See Channel

channel_egress_endpoint

See ChannelEgressEndpoint

channel_summary

See ChannelSummary

color_space_passthrough_settings

See ColorSpacePassthroughSettings

dvb_nit_settings

See DvbNitSettings

dvb_sdt_settings

See DvbSdtSettings

dvb_sub_destination_settings

See DvbSubDestinationSettings

dvb_sub_source_settings

See DvbSubSourceSettings

dvb_tdt_settings

See DvbTdtSettings

eac3_settings

See Eac3Settings

ebu_tt_d_destination_settings

See EbuTtDDestinationSettings

embedded_destination_settings

See EmbeddedDestinationSettings

embedded_plus_scte20_destination_settings

See EmbeddedPlusScte20DestinationSettings

embedded_source_settings

See EmbeddedSourceSettings

encoder_settings

See EncoderSettings

failover_condition

See FailoverCondition

failover_condition_settings

See FailoverConditionSettings

feature_activations

See FeatureActivations

fec_output_settings

See FecOutputSettings

fixed_mode_schedule_action_start_settings

See FixedModeScheduleActionStartSettings

fmp4_hls_settings

See Fmp4HlsSettings

follow_mode_schedule_action_start_settings

See FollowModeScheduleActionStartSettings

frame_capture_cdn_settings

See FrameCaptureCdnSettings

frame_capture_group_settings

See FrameCaptureGroupSettings

frame_capture_hls_settings

See FrameCaptureHlsSettings

frame_capture_output_settings

See FrameCaptureOutputSettings

frame_capture_s3_settings

See FrameCaptureS3Settings

frame_capture_settings

See FrameCaptureSettings

global_configuration

See GlobalConfiguration

h264_color_space_settings

See H264ColorSpaceSettings

h264_filter_settings

See H264FilterSettings

h264_settings

See H264Settings

h265_color_space_settings

See H265ColorSpaceSettings

h265_filter_settings

See H265FilterSettings

h265_settings

See H265Settings

hdr10_settings

See Hdr10Settings

hls_akamai_settings

See HlsAkamaiSettings

hls_basic_put_settings

See HlsBasicPutSettings

hls_cdn_settings

See HlsCdnSettings

hls_group_settings

See HlsGroupSettings

hls_id3_segment_tagging_schedule_action_settings

See HlsId3SegmentTaggingScheduleActionSettings

hls_input_settings

See HlsInputSettings

hls_media_store_settings

See HlsMediaStoreSettings

hls_output_settings

See HlsOutputSettings

hls_s3_settings

See HlsS3Settings

hls_settings

See HlsSettings

hls_timed_metadata_schedule_action_settings

See HlsTimedMetadataScheduleActionSettings

hls_webdav_settings

See HlsWebdavSettings

html_motion_graphics_settings

See HtmlMotionGraphicsSettings

immediate_mode_schedule_action_start_settings

See ImmediateModeScheduleActionStartSettings

input

See Input

input_attachment

See InputAttachment

input_channel_level

See InputChannelLevel

input_clipping_settings

See InputClippingSettings

input_destination

See InputDestination

input_destination_request

See InputDestinationRequest

input_destination_vpc

See InputDestinationVpc

input_device_configurable_settings

See InputDeviceConfigurableSettings

input_device_hd_settings

See InputDeviceHdSettings

input_device_network_settings

See InputDeviceNetworkSettings

input_device_request

See InputDeviceRequest

input_device_settings

See InputDeviceSettings

input_device_summary

See InputDeviceSummary

input_device_uhd_settings

See InputDeviceUhdSettings

input_location

See InputLocation

input_loss_behavior

See InputLossBehavior

input_loss_failover_settings

See InputLossFailoverSettings

input_prepare_schedule_action_settings

See InputPrepareScheduleActionSettings

input_security_group

See InputSecurityGroup

input_settings

See InputSettings

input_source

See InputSource

input_source_request

See InputSourceRequest

input_specification

See InputSpecification

input_switch_schedule_action_settings

See InputSwitchScheduleActionSettings

input_vpc_request

See InputVpcRequest

input_whitelist_rule

See InputWhitelistRule

input_whitelist_rule_cidr

See InputWhitelistRuleCidr

key_provider_settings

See KeyProviderSettings

m2ts_settings

See M2tsSettings

m3u8_settings

See M3u8Settings

media_connect_flow

See MediaConnectFlow

media_connect_flow_request

See MediaConnectFlowRequest

media_package_group_settings

See MediaPackageGroupSettings

media_package_output_destination_settings

See MediaPackageOutputDestinationSettings

media_package_output_settings

See MediaPackageOutputSettings

motion_graphics_activate_schedule_action_settings

See MotionGraphicsActivateScheduleActionSettings

motion_graphics_configuration

See MotionGraphicsConfiguration

motion_graphics_deactivate_schedule_action_settings

See MotionGraphicsDeactivateScheduleActionSettings

motion_graphics_settings

See MotionGraphicsSettings

mp2_settings

See Mp2Settings

mpeg2_filter_settings

See Mpeg2FilterSettings

mpeg2_settings

See Mpeg2Settings

ms_smooth_group_settings

See MsSmoothGroupSettings

ms_smooth_output_settings

See MsSmoothOutputSettings

multiplex

See Multiplex

multiplex_group_settings

See MultiplexGroupSettings

multiplex_media_connect_output_destination_settings

See MultiplexMediaConnectOutputDestinationSettings

multiplex_output_destination

See MultiplexOutputDestination

multiplex_output_settings

See MultiplexOutputSettings

multiplex_program

See MultiplexProgram

multiplex_program_channel_destination_settings

See MultiplexProgramChannelDestinationSettings

multiplex_program_packet_identifiers_map

See MultiplexProgramPacketIdentifiersMap

multiplex_program_pipeline_detail

See MultiplexProgramPipelineDetail

multiplex_program_service_descriptor

See MultiplexProgramServiceDescriptor

multiplex_program_settings

See MultiplexProgramSettings

multiplex_program_summary

See MultiplexProgramSummary

multiplex_settings

See MultiplexSettings

multiplex_settings_summary

See MultiplexSettingsSummary

multiplex_statmux_video_settings

See MultiplexStatmuxVideoSettings

multiplex_summary

See MultiplexSummary

multiplex_video_settings

See MultiplexVideoSettings

network_input_settings

See NetworkInputSettings

nielsen_configuration

See NielsenConfiguration

offering

See Offering

output

See Output

output_destination

See OutputDestination

output_destination_settings

See OutputDestinationSettings

output_group

See OutputGroup

output_group_settings

See OutputGroupSettings

output_location_ref

See OutputLocationRef

output_settings

See OutputSettings

pass_through_settings

See PassThroughSettings

pause_state_schedule_action_settings

See PauseStateScheduleActionSettings

pipeline_detail

See PipelineDetail

pipeline_pause_state_settings

See PipelinePauseStateSettings

raw_settings

See RawSettings

rec601_settings

See Rec601Settings

rec709_settings

See Rec709Settings

remix_settings

See RemixSettings

reservation

See Reservation

reservation_resource_specification

See ReservationResourceSpecification

rtmp_caption_info_destination_settings

See RtmpCaptionInfoDestinationSettings

rtmp_group_settings

See RtmpGroupSettings

rtmp_output_settings

See RtmpOutputSettings

schedule_action

See ScheduleAction

schedule_action_settings

See ScheduleActionSettings

schedule_action_start_settings

See ScheduleActionStartSettings

scte20_plus_embedded_destination_settings

See Scte20PlusEmbeddedDestinationSettings

scte20_source_settings

See Scte20SourceSettings

scte27_destination_settings

See Scte27DestinationSettings

scte27_source_settings

See Scte27SourceSettings

scte35_delivery_restrictions

See Scte35DeliveryRestrictions

scte35_descriptor

See Scte35Descriptor

scte35_descriptor_settings

See Scte35DescriptorSettings

scte35_return_to_network_schedule_action_settings

See Scte35ReturnToNetworkScheduleActionSettings

scte35_segmentation_descriptor

See Scte35SegmentationDescriptor

scte35_splice_insert

See Scte35SpliceInsert

scte35_splice_insert_schedule_action_settings

See Scte35SpliceInsertScheduleActionSettings

scte35_time_signal_apos

See Scte35TimeSignalApos

scte35_time_signal_schedule_action_settings

See Scte35TimeSignalScheduleActionSettings

smpte_tt_destination_settings

See SmpteTtDestinationSettings

standard_hls_settings

See StandardHlsSettings

start_timecode

See StartTimecode

static_image_activate_schedule_action_settings

See StaticImageActivateScheduleActionSettings

static_image_deactivate_schedule_action_settings

See StaticImageDeactivateScheduleActionSettings

static_key_settings

See StaticKeySettings

stop_timecode

See StopTimecode

teletext_destination_settings

See TeletextDestinationSettings

teletext_source_settings

See TeletextSourceSettings

temporal_filter_settings

See TemporalFilterSettings

timecode_config

See TimecodeConfig

transferring_input_device_summary

See TransferringInputDeviceSummary

ttml_destination_settings

See TtmlDestinationSettings

udp_container_settings

See UdpContainerSettings

udp_group_settings

See UdpGroupSettings

udp_output_settings

See UdpOutputSettings

validation_error

See ValidationError

video_black_failover_settings

See VideoBlackFailoverSettings

video_codec_settings

See VideoCodecSettings

video_description

See VideoDescription

video_selector

See VideoSelector

video_selector_color_space_settings

See VideoSelectorColorSpaceSettings

video_selector_pid

See VideoSelectorPid

video_selector_program_id

See VideoSelectorProgramId

video_selector_settings

See VideoSelectorSettings

vpc_output_settings

See VpcOutputSettings

vpc_output_settings_description

See VpcOutputSettingsDescription

wav_settings

See WavSettings

webvtt_destination_settings

See WebvttDestinationSettings

Structs

AacSettings

Aac Settings

Ac3Settings

Ac3 Settings

AncillarySourceSettings

Ancillary Source Settings

ArchiveCdnSettings

Archive Cdn Settings

ArchiveContainerSettings

Archive Container Settings

ArchiveGroupSettings

Archive Group Settings

ArchiveOutputSettings

Archive Output Settings

ArchiveS3Settings

Archive S3 Settings

AribDestinationSettings

Arib Destination Settings

AribSourceSettings

Arib Source Settings

AudioChannelMapping

Audio Channel Mapping

AudioCodecSettings

Audio Codec Settings

AudioDescription

Audio Description

AudioLanguageSelection

Audio Language Selection

AudioNormalizationSettings

Audio Normalization Settings

AudioOnlyHlsSettings

Audio Only Hls Settings

AudioPidSelection

Audio Pid Selection

AudioSelector

Audio Selector

AudioSelectorSettings

Audio Selector Settings

AudioSilenceFailoverSettings

Placeholder documentation for AudioSilenceFailoverSettings

AudioTrack

Audio Track

AudioTrackSelection

Audio Track Selection

AutomaticInputFailoverSettings

The settings for Automatic Input Failover.

AvailBlanking

Avail Blanking

AvailConfiguration

Avail Configuration

AvailSettings

Avail Settings

BatchFailedResultModel

Details from a failed operation

BatchScheduleActionCreateRequest

A list of schedule actions to create (in a request) or that have been created (in a response).

BatchScheduleActionCreateResult

List of actions that have been created in the schedule.

BatchScheduleActionDeleteRequest

A list of schedule actions to delete.

BatchScheduleActionDeleteResult

List of actions that have been deleted from the schedule.

BatchSuccessfulResultModel

Details from a successful operation

BlackoutSlate

Blackout Slate

BurnInDestinationSettings

Burn In Destination Settings

CaptionDescription

Caption Description

CaptionDestinationSettings

Caption Destination Settings

CaptionLanguageMapping

Maps a caption channel to an ISO 693-2 language code (http://www.loc.gov/standards/iso639-2), with an optional description.

CaptionRectangle

Caption Rectangle

CaptionSelector

Output groups for this Live Event. Output groups contain information about where streams should be distributed.

CaptionSelectorSettings

Caption Selector Settings

CdiInputSpecification

Placeholder documentation for CdiInputSpecification

Channel

Placeholder documentation for Channel

ChannelEgressEndpoint

Placeholder documentation for ChannelEgressEndpoint

ChannelSummary

Placeholder documentation for ChannelSummary

ColorSpacePassthroughSettings

Passthrough applies no color space conversion to the output

DvbNitSettings

DVB Network Information Table (NIT)

DvbSdtSettings

DVB Service Description Table (SDT)

DvbSubDestinationSettings

Dvb Sub Destination Settings

DvbSubSourceSettings

Dvb Sub Source Settings

DvbTdtSettings

DVB Time and Date Table (SDT)

Eac3Settings

Eac3 Settings

EbuTtDDestinationSettings

Ebu Tt DDestination Settings

EmbeddedDestinationSettings

Embedded Destination Settings

EmbeddedPlusScte20DestinationSettings

Embedded Plus Scte20 Destination Settings

EmbeddedSourceSettings

Embedded Source Settings

EncoderSettings

Encoder Settings

FailoverCondition

Failover Condition settings. There can be multiple failover conditions inside AutomaticInputFailoverSettings.

FailoverConditionSettings

Settings for one failover condition.

FeatureActivations

Feature Activations

FecOutputSettings

Fec Output Settings

FixedModeScheduleActionStartSettings

Start time for the action.

Fmp4HlsSettings

Fmp4 Hls Settings

FollowModeScheduleActionStartSettings

Settings to specify if an action follows another.

FrameCaptureCdnSettings

Frame Capture Cdn Settings

FrameCaptureGroupSettings

Frame Capture Group Settings

FrameCaptureHlsSettings

Frame Capture Hls Settings

FrameCaptureOutputSettings

Frame Capture Output Settings

FrameCaptureS3Settings

Frame Capture S3 Settings

FrameCaptureSettings

Frame Capture Settings

GlobalConfiguration

Global Configuration

H264ColorSpaceSettings

H264 Color Space Settings

H264FilterSettings

H264 Filter Settings

H264Settings

H264 Settings

H265ColorSpaceSettings

H265 Color Space Settings

H265FilterSettings

H265 Filter Settings

H265Settings

H265 Settings

Hdr10Settings

Hdr10 Settings

HlsAkamaiSettings

Hls Akamai Settings

HlsBasicPutSettings

Hls Basic Put Settings

HlsCdnSettings

Hls Cdn Settings

HlsGroupSettings

Hls Group Settings

HlsId3SegmentTaggingScheduleActionSettings

Settings for the action to insert a user-defined ID3 tag in each HLS segment

HlsInputSettings

Hls Input Settings

HlsMediaStoreSettings

Hls Media Store Settings

HlsOutputSettings

Hls Output Settings

HlsS3Settings

Hls S3 Settings

HlsSettings

Hls Settings

HlsTimedMetadataScheduleActionSettings

Settings for the action to emit HLS metadata

HlsWebdavSettings

Hls Webdav Settings

HtmlMotionGraphicsSettings

Html Motion Graphics Settings

ImmediateModeScheduleActionStartSettings

Settings to configure an action so that it occurs as soon as possible.

Input

Placeholder documentation for Input

InputAttachment

Placeholder documentation for InputAttachment

InputChannelLevel

Input Channel Level

InputClippingSettings

Settings to let you create a clip of the file input, in order to set up the input to ingest only a portion of the file.

InputDestination

The settings for a PUSH type input.

InputDestinationRequest

Endpoint settings for a PUSH type input.

InputDestinationVpc

The properties for a VPC type input destination.

InputDeviceConfigurableSettings

Configurable settings for the input device.

InputDeviceHdSettings

Settings that describe the active source from the input device, and the video characteristics of that source.

InputDeviceNetworkSettings

The network settings for the input device.

InputDeviceRequest

Settings for an input device.

InputDeviceSettings

Settings for an input device.

InputDeviceSummary

Details of the input device.

InputDeviceUhdSettings

Settings that describe the active source from the input device, and the video characteristics of that source.

InputLocation

Input Location

InputLossBehavior

Input Loss Behavior

InputLossFailoverSettings

MediaLive will perform a failover if content is not detected in this input for the specified period.

InputPrepareScheduleActionSettings

Action to prepare an input for a future immediate input switch.

InputSecurityGroup

An Input Security Group

InputSettings

Live Event input parameters. There can be multiple inputs in a single Live Event.

InputSource

The settings for a PULL type input.

InputSourceRequest

Settings for for a PULL type input.

InputSpecification

Placeholder documentation for InputSpecification

InputSwitchScheduleActionSettings

Settings for the “switch input” action: to switch from ingesting one input to ingesting another input.

InputVpcRequest

Settings for a private VPC Input. When this property is specified, the input destination addresses will be created in a VPC rather than with public Internet addresses. This property requires setting the roleArn property on Input creation. Not compatible with the inputSecurityGroups property.

InputWhitelistRule

Whitelist rule

InputWhitelistRuleCidr

An IPv4 CIDR to whitelist.

KeyProviderSettings

Key Provider Settings

M2tsSettings

M2ts Settings

M3u8Settings

Settings information for the .m3u8 container

MediaConnectFlow

The settings for a MediaConnect Flow.

MediaConnectFlowRequest

The settings for a MediaConnect Flow.

MediaPackageGroupSettings

Media Package Group Settings

MediaPackageOutputDestinationSettings

MediaPackage Output Destination Settings

MediaPackageOutputSettings

Media Package Output Settings

MotionGraphicsActivateScheduleActionSettings

Settings to specify the rendering of motion graphics into the video stream.

MotionGraphicsConfiguration

Motion Graphics Configuration

MotionGraphicsDeactivateScheduleActionSettings

Settings to specify the ending of rendering motion graphics into the video stream.

MotionGraphicsSettings

Motion Graphics Settings

Mp2Settings

Mp2 Settings

Mpeg2FilterSettings

Mpeg2 Filter Settings

Mpeg2Settings

Mpeg2 Settings

MsSmoothGroupSettings

Ms Smooth Group Settings

MsSmoothOutputSettings

Ms Smooth Output Settings

Multiplex

The multiplex object.

MultiplexGroupSettings

Multiplex Group Settings

MultiplexMediaConnectOutputDestinationSettings

Multiplex MediaConnect output destination settings.

MultiplexOutputDestination

Multiplex output destination settings

MultiplexOutputSettings

Multiplex Output Settings

MultiplexProgram

The multiplex program object.

MultiplexProgramChannelDestinationSettings

Multiplex Program Input Destination Settings for outputting a Channel to a Multiplex

MultiplexProgramPacketIdentifiersMap

Packet identifiers map for a given Multiplex program.

MultiplexProgramPipelineDetail

The current source for one of the pipelines in the multiplex.

MultiplexProgramServiceDescriptor

Transport stream service descriptor configuration for the Multiplex program.

MultiplexProgramSettings

Multiplex Program settings configuration.

MultiplexProgramSummary

Placeholder documentation for MultiplexProgramSummary

MultiplexSettings

Contains configuration for a Multiplex event

MultiplexSettingsSummary

Contains summary configuration for a Multiplex event.

MultiplexStatmuxVideoSettings

Statmux rate control settings

MultiplexSummary

Placeholder documentation for MultiplexSummary

MultiplexVideoSettings

The video configuration for each program in a multiplex.

NetworkInputSettings

Network source to transcode. Must be accessible to the Elemental Live node that is running the live event through a network connection.

NielsenConfiguration

Nielsen Configuration

Offering

Reserved resources available for purchase

Output

Output settings. There can be multiple outputs within a group.

OutputDestination

Placeholder documentation for OutputDestination

OutputDestinationSettings

Placeholder documentation for OutputDestinationSettings

OutputGroup

Output groups for this Live Event. Output groups contain information about where streams should be distributed.

OutputGroupSettings

Output Group Settings

OutputLocationRef

Reference to an OutputDestination ID defined in the channel

OutputSettings

Output Settings

PassThroughSettings

Pass Through Settings

PauseStateScheduleActionSettings

Settings for the action to set pause state of a channel.

PipelineDetail

Runtime details of a pipeline when a channel is running.

PipelinePauseStateSettings

Settings for pausing a pipeline.

RawSettings

Raw Settings

Rec601Settings

Rec601 Settings

Rec709Settings

Rec709 Settings

RemixSettings

Remix Settings

Reservation

Reserved resources available to use

ReservationResourceSpecification

Resource configuration (codec, resolution, bitrate, …)

RtmpCaptionInfoDestinationSettings

Rtmp Caption Info Destination Settings

RtmpGroupSettings

Rtmp Group Settings

RtmpOutputSettings

Rtmp Output Settings

ScheduleAction

Contains information on a single schedule action.

ScheduleActionSettings

Holds the settings for a single schedule action.

ScheduleActionStartSettings

Settings to specify when an action should occur. Only one of the options must be selected.

Scte20PlusEmbeddedDestinationSettings

Scte20 Plus Embedded Destination Settings

Scte20SourceSettings

Scte20 Source Settings

Scte27DestinationSettings

Scte27 Destination Settings

Scte27SourceSettings

Scte27 Source Settings

Scte35DeliveryRestrictions

Corresponds to SCTE-35 delivery_not_restricted_flag parameter. To declare delivery restrictions, include this element and its four “restriction” flags. To declare that there are no restrictions, omit this element.

Scte35Descriptor

Holds one set of SCTE-35 Descriptor Settings.

Scte35DescriptorSettings

SCTE-35 Descriptor settings.

Scte35ReturnToNetworkScheduleActionSettings

Settings for a SCTE-35 return_to_network message.

Scte35SegmentationDescriptor

Corresponds to SCTE-35 segmentation_descriptor.

Scte35SpliceInsert

Scte35 Splice Insert

Scte35SpliceInsertScheduleActionSettings

Settings for a SCTE-35 splice_insert message.

Scte35TimeSignalApos

Scte35 Time Signal Apos

Scte35TimeSignalScheduleActionSettings

Settings for a SCTE-35 time_signal.

SmpteTtDestinationSettings

Smpte Tt Destination Settings

StandardHlsSettings

Standard Hls Settings

StartTimecode

Settings to identify the start of the clip.

StaticImageActivateScheduleActionSettings

Settings for the action to activate a static image.

StaticImageDeactivateScheduleActionSettings

Settings for the action to deactivate the image in a specific layer.

StaticKeySettings

Static Key Settings

StopTimecode

Settings to identify the end of the clip.

TeletextDestinationSettings

Teletext Destination Settings

TeletextSourceSettings

Teletext Source Settings

TemporalFilterSettings

Temporal Filter Settings

TimecodeConfig

Timecode Config

TransferringInputDeviceSummary

Details about the input device that is being transferred.

TtmlDestinationSettings

Ttml Destination Settings

UdpContainerSettings

Udp Container Settings

UdpGroupSettings

Udp Group Settings

UdpOutputSettings

Udp Output Settings

ValidationError

Placeholder documentation for ValidationError

VideoBlackFailoverSettings

Placeholder documentation for VideoBlackFailoverSettings

VideoCodecSettings

Video Codec Settings

VideoDescription

Video settings for this stream.

VideoSelector

Specifies a particular video stream within an input source. An input may have only a single video selector.

VideoSelectorColorSpaceSettings

Video Selector Color Space Settings

VideoSelectorPid

Video Selector Pid

VideoSelectorProgramId

Video Selector Program Id

VideoSelectorSettings

Video Selector Settings

VpcOutputSettings

The properties for a private VPC Output When this property is specified, the output egress addresses will be created in a user specified VPC

VpcOutputSettingsDescription

The properties for a private VPC Output

WavSettings

Wav Settings

WebvttDestinationSettings

Webvtt Destination Settings

Enums

AacCodingMode

Aac Coding Mode

AacInputType

Aac Input Type

AacProfile

Aac Profile

AacRateControlMode

Aac Rate Control Mode

AacRawFormat

Aac Raw Format

AacSpec

Aac Spec

AacVbrQuality

Aac Vbr Quality

Ac3BitstreamMode

Ac3 Bitstream Mode

Ac3CodingMode

Ac3 Coding Mode

Ac3DrcProfile

Ac3 Drc Profile

Ac3LfeFilter

Ac3 Lfe Filter

Ac3MetadataControl

Ac3 Metadata Control

AcceptHeader

The HTTP Accept header. Indicates the requested type fothe thumbnail.

AfdSignaling

Afd Signaling

AudioDescriptionAudioTypeControl

Audio Description Audio Type Control

AudioDescriptionLanguageCodeControl

Audio Description Language Code Control

AudioLanguageSelectionPolicy

Audio Language Selection Policy

AudioNormalizationAlgorithm

Audio Normalization Algorithm

AudioNormalizationAlgorithmControl

Audio Normalization Algorithm Control

AudioOnlyHlsSegmentType

Audio Only Hls Segment Type

AudioOnlyHlsTrackType

Audio Only Hls Track Type

AudioType

Audio Type

AuthenticationScheme

Authentication Scheme

AvailBlankingState

Avail Blanking State

BlackoutSlateNetworkEndBlackout

Blackout Slate Network End Blackout

BlackoutSlateState

Blackout Slate State

BurnInAlignment

Burn In Alignment

BurnInBackgroundColor

Burn In Background Color

BurnInFontColor

Burn In Font Color

BurnInOutlineColor

Burn In Outline Color

BurnInShadowColor

Burn In Shadow Color

BurnInTeletextGridControl

Burn In Teletext Grid Control

CdiInputResolution

Maximum CDI input resolution; SD is 480i and 576i up to 30 frames-per-second (fps), HD is 720p up to 60 fps / 1080i up to 30 fps, FHD is 1080p up to 60 fps, UHD is 2160p up to 60 fps

ChannelClass

A standard channel has two encoding pipelines and a single pipeline channel only has one.

ChannelState

Placeholder documentation for ChannelState

ContentType

Specifies the media type of the thumbnail.

DeviceSettingsSyncState

The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

DeviceUpdateStatus

The status of software on the input device.

DvbSdtOutputSdt

Dvb Sdt Output Sdt

DvbSubDestinationAlignment

Dvb Sub Destination Alignment

DvbSubDestinationBackgroundColor

Dvb Sub Destination Background Color

DvbSubDestinationFontColor

Dvb Sub Destination Font Color

DvbSubDestinationOutlineColor

Dvb Sub Destination Outline Color

DvbSubDestinationShadowColor

Dvb Sub Destination Shadow Color

DvbSubDestinationTeletextGridControl

Dvb Sub Destination Teletext Grid Control

DvbSubOcrLanguage

Dvb Sub Ocr Language

Eac3AttenuationControl

Eac3 Attenuation Control

Eac3BitstreamMode

Eac3 Bitstream Mode

Eac3CodingMode

Eac3 Coding Mode

Eac3DcFilter

Eac3 Dc Filter

Eac3DrcLine

Eac3 Drc Line

Eac3DrcRf

Eac3 Drc Rf

Eac3LfeControl

Eac3 Lfe Control

Eac3LfeFilter

Eac3 Lfe Filter

Eac3MetadataControl

Eac3 Metadata Control

Eac3PassthroughControl

Eac3 Passthrough Control

Eac3PhaseControl

Eac3 Phase Control

Eac3StereoDownmix

Eac3 Stereo Downmix

Eac3SurroundExMode

Eac3 Surround Ex Mode

Eac3SurroundMode

Eac3 Surround Mode

EbuTtDDestinationStyleControl

Ebu Tt DDestination Style Control

EbuTtDFillLineGapControl

Ebu Tt DFill Line Gap Control

EmbeddedConvert608To708

Embedded Convert608 To708

EmbeddedScte20Detection

Embedded Scte20 Detection

FeatureActivationsInputPrepareScheduleActions

Feature Activations Input Prepare Schedule Actions

FecOutputIncludeFec

Fec Output Include Fec

FixedAfd

Fixed Afd

Fmp4NielsenId3Behavior

Fmp4 Nielsen Id3 Behavior

Fmp4TimedMetadataBehavior

Fmp4 Timed Metadata Behavior

FollowPoint

Follow reference point.

FrameCaptureIntervalUnit

Frame Capture Interval Unit

GlobalConfigurationInputEndAction

Global Configuration Input End Action

GlobalConfigurationLowFramerateInputs

Global Configuration Low Framerate Inputs

GlobalConfigurationOutputLockingMode

Global Configuration Output Locking Mode

GlobalConfigurationOutputTimingSource

Global Configuration Output Timing Source

H264AdaptiveQuantization

H264 Adaptive Quantization

H264ColorMetadata

H264 Color Metadata

H264EntropyEncoding

H264 Entropy Encoding

H264FlickerAq

H264 Flicker Aq

H264ForceFieldPictures

H264 Force Field Pictures

H264FramerateControl

H264 Framerate Control

H264GopBReference

H264 Gop BReference

H264GopSizeUnits

H264 Gop Size Units

H264Level

H264 Level

H264LookAheadRateControl

H264 Look Ahead Rate Control

H264ParControl

H264 Par Control

H264Profile

H264 Profile

H264QualityLevel

H264 Quality Level

H264RateControlMode

H264 Rate Control Mode

H264ScanType

H264 Scan Type

H264SceneChangeDetect

H264 Scene Change Detect

H264SpatialAq

H264 Spatial Aq

H264SubGopLength

H264 Sub Gop Length

H264Syntax

H264 Syntax

H264TemporalAq

H264 Temporal Aq

H264TimecodeInsertionBehavior

H264 Timecode Insertion Behavior

H265AdaptiveQuantization

H265 Adaptive Quantization

H265AlternativeTransferFunction

H265 Alternative Transfer Function

H265ColorMetadata

H265 Color Metadata

H265FlickerAq

H265 Flicker Aq

H265GopSizeUnits

H265 Gop Size Units

H265Level

H265 Level

H265LookAheadRateControl

H265 Look Ahead Rate Control

H265Profile

H265 Profile

H265RateControlMode

H265 Rate Control Mode

H265ScanType

H265 Scan Type

H265SceneChangeDetect

H265 Scene Change Detect

H265Tier

H265 Tier

H265TimecodeInsertionBehavior

H265 Timecode Insertion Behavior

HlsAdMarkers

Hls Ad Markers

HlsAkamaiHttpTransferMode

Hls Akamai Http Transfer Mode

HlsCaptionLanguageSetting

Hls Caption Language Setting

HlsClientCache

Hls Client Cache

HlsCodecSpecification

Hls Codec Specification

HlsDirectoryStructure

Hls Directory Structure

HlsDiscontinuityTags

Hls Discontinuity Tags

HlsEncryptionType

Hls Encryption Type

HlsH265PackagingType

Hls H265 Packaging Type

HlsId3SegmentTaggingState

State of HLS ID3 Segment Tagging

HlsIncompleteSegmentBehavior

Hls Incomplete Segment Behavior

HlsIvInManifest

Hls Iv In Manifest

HlsIvSource

Hls Iv Source

HlsManifestCompression

Hls Manifest Compression

HlsManifestDurationFormat

Hls Manifest Duration Format

HlsMediaStoreStorageClass

Hls Media Store Storage Class

HlsMode

Hls Mode

HlsOutputSelection

Hls Output Selection

HlsProgramDateTime

Hls Program Date Time

HlsRedundantManifest

Hls Redundant Manifest

HlsScte35SourceType

Hls Scte35 Source Type

HlsSegmentationMode

Hls Segmentation Mode

HlsStreamInfResolution

Hls Stream Inf Resolution

HlsTimedMetadataId3Frame

Hls Timed Metadata Id3 Frame

HlsTsFileMode

Hls Ts File Mode

HlsWebdavHttpTransferMode

Hls Webdav Http Transfer Mode

IFrameOnlyPlaylistType

When set to “standard”, an I-Frame only playlist will be written out for each video output in the output group. This I-Frame only playlist will contain byte range offsets pointing to the I-frame(s) in each segment.

InputClass

A standard input has two sources and a single pipeline input only has one.

InputCodec

codec in increasing order of complexity

InputDeblockFilter

Input Deblock Filter

InputDenoiseFilter

Input Denoise Filter

InputDeviceActiveInput

The source at the input device that is currently active.

InputDeviceConfiguredInput

The source to activate (use) from the input device.

InputDeviceConnectionState

The state of the connection between the input device and AWS.

InputDeviceIpScheme

Specifies whether the input device has been configured (outside of MediaLive) to use a dynamic IP address assignment (DHCP) or a static IP address.

InputDeviceScanType

The scan type of the video source.

InputDeviceState

The state of the input device.

InputDeviceTransferType

The type of device transfer. INCOMING for an input device that is being transferred to you, OUTGOING for an input device that you are transferring to another AWS account.

InputDeviceType

The type of the input device. For an AWS Elemental Link device that outputs resolutions up to 1080, choose “HD”.

InputFilter

Input Filter

InputLossActionForHlsOut

Input Loss Action For Hls Out

InputLossActionForMsSmoothOut

Input Loss Action For Ms Smooth Out

InputLossActionForRtmpOut

Input Loss Action For Rtmp Out

InputLossActionForUdpOut

Input Loss Action For Udp Out

InputLossImageType

Input Loss Image Type

InputMaximumBitrate

Maximum input bitrate in megabits per second. Bitrates up to 50 Mbps are supported currently.

InputPreference

Input preference when deciding which input to make active when a previously failed input has recovered. If "EQUAL_INPUT_PREFERENCE", then the active input will stay active as long as it is healthy. If "PRIMARY_INPUT_PREFERRED", then always switch back to the primary input when it is healthy.

InputResolution

Input resolution based on lines of vertical resolution in the input; SD is less than 720 lines, HD is 720 to 1080 lines, UHD is greater than 1080 lines

InputSecurityGroupState

Placeholder documentation for InputSecurityGroupState

InputSourceEndBehavior

Input Source End Behavior

InputSourceType

There are two types of input sources, static and dynamic. If an input source is dynamic you can change the source url of the input dynamically using an input switch action. However, the only input type to support a dynamic url at this time is MP4_FILE. By default all input sources are static.

InputState

Placeholder documentation for InputState

InputTimecodeSource

Documentation update needed

InputType

Placeholder documentation for InputType

LastFrameClippingBehavior

If you specify a StopTimecode in an input (in order to clip the file), you can specify if you want the clip to exclude (the default) or include the frame specified by the timecode.

LogLevel

The log level the user wants for their channel.

M2tsAbsentInputAudioBehavior

M2ts Absent Input Audio Behavior

M2tsArib

M2ts Arib

M2tsAribCaptionsPidControl

M2ts Arib Captions Pid Control

M2tsAudioBufferModel

M2ts Audio Buffer Model

M2tsAudioInterval

M2ts Audio Interval

M2tsAudioStreamType

M2ts Audio Stream Type

M2tsBufferModel

M2ts Buffer Model

M2tsCcDescriptor

M2ts Cc Descriptor

M2tsEbifControl

M2ts Ebif Control

M2tsEbpPlacement

M2ts Ebp Placement

M2tsEsRateInPes

M2ts Es Rate In Pes

M2tsKlv

M2ts Klv

M2tsNielsenId3Behavior

M2ts Nielsen Id3 Behavior

M2tsPcrControl

M2ts Pcr Control

M2tsRateMode

M2ts Rate Mode

M2tsScte35Control

M2ts Scte35 Control

M2tsSegmentationMarkers

M2ts Segmentation Markers

M2tsSegmentationStyle

M2ts Segmentation Style

M2tsTimedMetadataBehavior

M2ts Timed Metadata Behavior

M3u8NielsenId3Behavior

M3u8 Nielsen Id3 Behavior

M3u8PcrControl

M3u8 Pcr Control

M3u8Scte35Behavior

M3u8 Scte35 Behavior

M3u8TimedMetadataBehavior

M3u8 Timed Metadata Behavior

MotionGraphicsInsertion

Motion Graphics Insertion

Mp2CodingMode

Mp2 Coding Mode

Mpeg2AdaptiveQuantization

Mpeg2 Adaptive Quantization

Mpeg2ColorMetadata

Mpeg2 Color Metadata

Mpeg2ColorSpace

Mpeg2 Color Space

Mpeg2DisplayRatio

Mpeg2 Display Ratio

Mpeg2GopSizeUnits

Mpeg2 Gop Size Units

Mpeg2ScanType

Mpeg2 Scan Type

Mpeg2SubGopLength

Mpeg2 Sub Gop Length

Mpeg2TimecodeInsertionBehavior

Mpeg2 Timecode Insertion Behavior

MsSmoothH265PackagingType

Ms Smooth H265 Packaging Type

MultiplexState

The current state of the multiplex.

NetworkInputServerValidation

Network Input Server Validation

NielsenPcmToId3TaggingState

State of Nielsen PCM to ID3 tagging

OfferingDurationUnits

Units for duration, e.g. ‘MONTHS’

OfferingType

Offering type, e.g. ‘NO_UPFRONT’

PipelineId

Pipeline ID

PreferredChannelPipeline

Indicates which pipeline is preferred by the multiplex for program ingest. If set to "PIPELINE_0" or "PIPELINE_1" and an unhealthy ingest causes the multiplex to switch to the non-preferred pipeline, it will switch back once that ingest is healthy again. If set to "CURRENTLY_ACTIVE", it will not switch back to the other pipeline based on it recovering to a healthy state, it will only switch if the active pipeline becomes unhealthy.

ReservationCodec

Codec, ‘MPEG2’, ‘AVC’, ‘HEVC’, or ‘AUDIO’

ReservationMaximumBitrate

Maximum bitrate in megabits per second

ReservationMaximumFramerate

Maximum framerate in frames per second (Outputs only)

ReservationResolution

Resolution based on lines of vertical resolution; SD is less than 720 lines, HD is 720 to 1080 lines, FHD is 1080 lines, UHD is greater than 1080 lines

ReservationResourceType

Resource type, ‘INPUT’, ‘OUTPUT’, ‘MULTIPLEX’, or ‘CHANNEL’

ReservationSpecialFeature

Special features, ‘ADVANCED_AUDIO’ ‘AUDIO_NORMALIZATION’ ‘MGHD’ or ‘MGUHD’

ReservationState

Current reservation state

ReservationVideoQuality

Video quality, e.g. ‘STANDARD’ (Outputs only)

RtmpAdMarkers

Rtmp Ad Markers

RtmpCacheFullBehavior

Rtmp Cache Full Behavior

RtmpCaptionData

Rtmp Caption Data

RtmpOutputCertificateMode

Rtmp Output Certificate Mode

S3CannedAcl

S3 Canned Acl

Scte20Convert608To708

Scte20 Convert608 To708

Scte27OcrLanguage

Scte27 Ocr Language

Scte35AposNoRegionalBlackoutBehavior

Scte35 Apos No Regional Blackout Behavior

Scte35AposWebDeliveryAllowedBehavior

Scte35 Apos Web Delivery Allowed Behavior

Scte35ArchiveAllowedFlag

Corresponds to the archive_allowed parameter. A value of ARCHIVE_NOT_ALLOWED corresponds to 0 (false) in the SCTE-35 specification. If you include one of the “restriction” flags then you must include all four of them.

Scte35DeviceRestrictions

Corresponds to the device_restrictions parameter in a segmentation_descriptor. If you include one of the “restriction” flags then you must include all four of them.

Scte35NoRegionalBlackoutFlag

Corresponds to the no_regional_blackout_flag parameter. A value of REGIONAL_BLACKOUT corresponds to 0 (false) in the SCTE-35 specification. If you include one of the “restriction” flags then you must include all four of them.

Scte35SegmentationCancelIndicator

Corresponds to SCTE-35 segmentation_event_cancel_indicator. SEGMENTATION_EVENT_NOT_CANCELED corresponds to 0 in the SCTE-35 specification and indicates that this is an insertion request. SEGMENTATION_EVENT_CANCELED corresponds to 1 in the SCTE-35 specification and indicates that this is a cancelation request, in which case complete this field and the existing event ID to cancel.

Scte35SpliceInsertNoRegionalBlackoutBehavior

Scte35 Splice Insert No Regional Blackout Behavior

Scte35SpliceInsertWebDeliveryAllowedBehavior

Scte35 Splice Insert Web Delivery Allowed Behavior

Scte35WebDeliveryAllowedFlag

Corresponds to the web_delivery_allowed_flag parameter. A value of WEB_DELIVERY_NOT_ALLOWED corresponds to 0 (false) in the SCTE-35 specification. If you include one of the “restriction” flags then you must include all four of them.

SmoothGroupAudioOnlyTimecodeControl

Smooth Group Audio Only Timecode Control

SmoothGroupCertificateMode

Smooth Group Certificate Mode

SmoothGroupEventIdMode

Smooth Group Event Id Mode

SmoothGroupEventStopBehavior

Smooth Group Event Stop Behavior

SmoothGroupSegmentationMode

Smooth Group Segmentation Mode

SmoothGroupSparseTrackType

Smooth Group Sparse Track Type

SmoothGroupStreamManifestBehavior

Smooth Group Stream Manifest Behavior

SmoothGroupTimestampOffsetMode

Smooth Group Timestamp Offset Mode

Smpte2038DataPreference

Smpte2038 Data Preference

TemporalFilterPostFilterSharpening

Temporal Filter Post Filter Sharpening

TemporalFilterStrength

Temporal Filter Strength

TimecodeConfigSource

Timecode Config Source

TtmlDestinationStyleControl

Ttml Destination Style Control

UdpTimedMetadataId3Frame

Udp Timed Metadata Id3 Frame

VideoDescriptionRespondToAfd

Video Description Respond To Afd

VideoDescriptionScalingBehavior

Video Description Scaling Behavior

VideoSelectorColorSpace

Video Selector Color Space

VideoSelectorColorSpaceUsage

Video Selector Color Space Usage

WavCodingMode

Wav Coding Mode