Struct aws_sdk_medialive::model::dvb_sub_destination_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DvbSubDestinationSettings

Implementations

impl Builder

pub fn alignment(self, input: DvbSubDestinationAlignment) -> Self

If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting “smart” justification will left-justify live subtitles and center-justify pre-recorded subtitles. This option is not valid for source captions that are STL or 608/embedded. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_alignment(self, input: Option<DvbSubDestinationAlignment>) -> Self

pub fn background_color(self, input: DvbSubDestinationBackgroundColor) -> Self

Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

pub fn set_background_color(
self,
input: Option<DvbSubDestinationBackgroundColor>
) -> Self

pub fn background_opacity(self, input: i32) -> Self

Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

pub fn set_background_opacity(self, input: Option<i32>) -> Self

pub fn font(self, input: InputLocation) -> Self

External font file used for caption burn-in. File extension must be ‘ttf’ or ‘tte’. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

pub fn set_font(self, input: Option<InputLocation>) -> Self

pub fn font_color(self, input: DvbSubDestinationFontColor) -> Self

Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_font_color(self, input: Option<DvbSubDestinationFontColor>) -> Self

pub fn font_opacity(self, input: i32) -> Self

Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

pub fn set_font_opacity(self, input: Option<i32>) -> Self

pub fn font_resolution(self, input: i32) -> Self

Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

pub fn set_font_resolution(self, input: Option<i32>) -> Self

pub fn font_size(self, input: impl Into<String>) -> Self

When set to auto fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

pub fn set_font_size(self, input: Option<String>) -> Self

pub fn outline_color(self, input: DvbSubDestinationOutlineColor) -> Self

Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_outline_color(
self,
input: Option<DvbSubDestinationOutlineColor>
) -> Self

pub fn outline_size(self, input: i32) -> Self

Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_outline_size(self, input: Option<i32>) -> Self

pub fn shadow_color(self, input: DvbSubDestinationShadowColor) -> Self

Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

pub fn set_shadow_color(
self,
input: Option<DvbSubDestinationShadowColor>
) -> Self

pub fn shadow_opacity(self, input: i32) -> Self

Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

pub fn set_shadow_opacity(self, input: Option<i32>) -> Self

pub fn shadow_x_offset(self, input: i32) -> Self

Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

pub fn set_shadow_x_offset(self, input: Option<i32>) -> Self

pub fn shadow_y_offset(self, input: i32) -> Self

Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

pub fn set_shadow_y_offset(self, input: Option<i32>) -> Self

pub fn teletext_grid_control(
self,
input: DvbSubDestinationTeletextGridControl
) -> Self

Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

pub fn set_teletext_grid_control(
self,
input: Option<DvbSubDestinationTeletextGridControl>
) -> Self

pub fn x_position(self, input: i32) -> Self

Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_x_position(self, input: Option<i32>) -> Self

pub fn y_position(self, input: i32) -> Self

Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

pub fn set_y_position(self, input: Option<i32>) -> Self

pub fn build(self) -> DvbSubDestinationSettings

Consumes the builder and constructs a DvbSubDestinationSettings