Struct aws_sdk_medialive::model::h264_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for H264Settings

Implementations

impl Builder

pub fn adaptive_quantization(self, input: H264AdaptiveQuantization) -> Self

Enables or disables adaptive quantization, which is a technique MediaLive can apply to video on a frame-by-frame basis to produce more compression without losing quality. There are three types of adaptive quantization: flicker, spatial, and temporal. Set the field in one of these ways: Set to Auto. Recommended. For each type of AQ, MediaLive will determine if AQ is needed, and if so, the appropriate strength. Set a strength (a value other than Auto or Disable). This strength will apply to any of the AQ fields that you choose to enable. Set to Disabled to disable all types of adaptive quantization.

pub fn set_adaptive_quantization(
self,
input: Option<H264AdaptiveQuantization>
) -> Self

pub fn afd_signaling(self, input: AfdSignaling) -> Self

Indicates that AFD values will be written into the output stream. If afdSignaling is “auto”, the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to “fixed”, the AFD value will be the value configured in the fixedAfd parameter.

pub fn set_afd_signaling(self, input: Option<AfdSignaling>) -> Self

pub fn bitrate(self, input: i32) -> Self

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

pub fn set_bitrate(self, input: Option<i32>) -> Self

pub fn buf_fill_pct(self, input: i32) -> Self

Percentage of the buffer that should initially be filled (HRD buffer model).

pub fn set_buf_fill_pct(self, input: Option<i32>) -> Self

pub fn buf_size(self, input: i32) -> Self

Size of buffer (HRD buffer model) in bits.

pub fn set_buf_size(self, input: Option<i32>) -> Self

pub fn color_metadata(self, input: H264ColorMetadata) -> Self

Includes colorspace metadata in the output.

pub fn set_color_metadata(self, input: Option<H264ColorMetadata>) -> Self

pub fn color_space_settings(self, input: H264ColorSpaceSettings) -> Self

Color Space settings

pub fn set_color_space_settings(
self,
input: Option<H264ColorSpaceSettings>
) -> Self

pub fn entropy_encoding(self, input: H264EntropyEncoding) -> Self

Entropy encoding mode. Use cabac (must be in Main or High profile) or cavlc.

pub fn set_entropy_encoding(self, input: Option<H264EntropyEncoding>) -> Self

pub fn filter_settings(self, input: H264FilterSettings) -> Self

Optional filters that you can apply to an encode.

pub fn set_filter_settings(self, input: Option<H264FilterSettings>) -> Self

pub fn fixed_afd(self, input: FixedAfd) -> Self

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to ‘Fixed’.

pub fn set_fixed_afd(self, input: Option<FixedAfd>) -> Self

pub fn flicker_aq(self, input: H264FlickerAq) -> Self

Flicker AQ makes adjustments within each frame to reduce flicker or ‘pop’ on I-frames. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if flicker AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply flicker AQ using the specified strength. Disabled: MediaLive won’t apply flicker AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn’t apply flicker AQ.

pub fn set_flicker_aq(self, input: Option<H264FlickerAq>) -> Self

pub fn force_field_pictures(self, input: H264ForceFieldPictures) -> Self

This setting applies only when scan type is “interlaced.” It controls whether coding is performed on a field basis or on a frame basis. (When the video is progressive, the coding is always performed on a frame basis.) enabled: Force MediaLive to code on a field basis, so that odd and even sets of fields are coded separately. disabled: Code the two sets of fields separately (on a field basis) or together (on a frame basis using PAFF), depending on what is most appropriate for the content.

pub fn set_force_field_pictures(
self,
input: Option<H264ForceFieldPictures>
) -> Self

pub fn framerate_control(self, input: H264FramerateControl) -> Self

This field indicates how the output video frame rate is specified. If “specified” is selected then the output video frame rate is determined by framerateNumerator and framerateDenominator, else if “initializeFromSource” is selected then the output video frame rate will be set equal to the input video frame rate of the first input.

pub fn set_framerate_control(self, input: Option<H264FramerateControl>) -> Self

pub fn framerate_denominator(self, input: i32) -> Self

Framerate denominator.

pub fn set_framerate_denominator(self, input: Option<i32>) -> Self

pub fn framerate_numerator(self, input: i32) -> Self

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

pub fn set_framerate_numerator(self, input: Option<i32>) -> Self

pub fn gop_b_reference(self, input: H264GopBReference) -> Self

Documentation update needed

pub fn set_gop_b_reference(self, input: Option<H264GopBReference>) -> Self

pub fn gop_closed_cadence(self, input: i32) -> Self

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

pub fn set_gop_closed_cadence(self, input: Option<i32>) -> Self

pub fn gop_num_b_frames(self, input: i32) -> Self

Number of B-frames between reference frames.

pub fn set_gop_num_b_frames(self, input: Option<i32>) -> Self

pub fn gop_size(self, input: f64) -> Self

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

pub fn set_gop_size(self, input: Option<f64>) -> Self

pub fn gop_size_units(self, input: H264GopSizeUnits) -> Self

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

pub fn set_gop_size_units(self, input: Option<H264GopSizeUnits>) -> Self

pub fn level(self, input: H264Level) -> Self

H.264 Level.

pub fn set_level(self, input: Option<H264Level>) -> Self

pub fn look_ahead_rate_control(self, input: H264LookAheadRateControl) -> Self

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

pub fn set_look_ahead_rate_control(
self,
input: Option<H264LookAheadRateControl>
) -> Self

pub fn max_bitrate(self, input: i32) -> Self

For QVBR: See the tooltip for Quality level For VBR: Set the maximum bitrate in order to accommodate expected spikes in the complexity of the video.

pub fn set_max_bitrate(self, input: Option<i32>) -> Self

pub fn min_i_interval(self, input: i32) -> Self

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

pub fn set_min_i_interval(self, input: Option<i32>) -> Self

pub fn num_ref_frames(self, input: i32) -> Self

Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

pub fn set_num_ref_frames(self, input: Option<i32>) -> Self

pub fn par_control(self, input: H264ParControl) -> Self

This field indicates how the output pixel aspect ratio is specified. If “specified” is selected then the output video pixel aspect ratio is determined by parNumerator and parDenominator, else if “initializeFromSource” is selected then the output pixsel aspect ratio will be set equal to the input video pixel aspect ratio of the first input.

pub fn set_par_control(self, input: Option<H264ParControl>) -> Self

pub fn par_denominator(self, input: i32) -> Self

Pixel Aspect Ratio denominator.

pub fn set_par_denominator(self, input: Option<i32>) -> Self

pub fn par_numerator(self, input: i32) -> Self

Pixel Aspect Ratio numerator.

pub fn set_par_numerator(self, input: Option<i32>) -> Self

pub fn profile(self, input: H264Profile) -> Self

H.264 Profile.

pub fn set_profile(self, input: Option<H264Profile>) -> Self

pub fn quality_level(self, input: H264QualityLevel) -> Self

Leave as STANDARD_QUALITY or choose a different value (which might result in additional costs to run the channel).

ENHANCED_QUALITY: Produces a slightly better video quality without an increase in the bitrate. Has an effect only when the Rate control mode is QVBR or CBR. If this channel is in a MediaLive multiplex, the value must be ENHANCED_QUALITY.

STANDARD_QUALITY: Valid for any Rate control mode.

pub fn set_quality_level(self, input: Option<H264QualityLevel>) -> Self

pub fn qvbr_quality_level(self, input: i32) -> Self

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. You can set a target quality or you can let MediaLive determine the best quality. To set a target quality, enter values in the QVBR quality level field and the Max bitrate field. Enter values that suit your most important viewing devices. Recommended values are:

Primary screen: Quality level: 8 to 10. Max bitrate: 4M

PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M

Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M To let MediaLive decide, leave the QVBR quality level field empty, and in Max bitrate enter the maximum rate you want in the video. For more information, see the section called “Video - rate control mode” in the MediaLive user guide

pub fn set_qvbr_quality_level(self, input: Option<i32>) -> Self

pub fn rate_control_mode(self, input: H264RateControlMode) -> Self

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. VBR: Quality and bitrate vary, depending on the video complexity. Recommended instead of QVBR if you want to maintain a specific average bitrate over the duration of the channel. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

pub fn set_rate_control_mode(self, input: Option<H264RateControlMode>) -> Self

pub fn scan_type(self, input: H264ScanType) -> Self

Sets the scan type of the output to progressive or top-field-first interlaced.

pub fn set_scan_type(self, input: Option<H264ScanType>) -> Self

pub fn scene_change_detect(self, input: H264SceneChangeDetect) -> Self

Scene change detection.

On: inserts I-frames when scene change is detected.

Off: does not force an I-frame when scene change is detected.

pub fn set_scene_change_detect(
self,
input: Option<H264SceneChangeDetect>
) -> Self

pub fn slices(self, input: i32) -> Self

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

pub fn set_slices(self, input: Option<i32>) -> Self

pub fn softness(self, input: i32) -> Self

Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image. If not set to zero, must be greater than 15.

pub fn set_softness(self, input: Option<i32>) -> Self

pub fn spatial_aq(self, input: H264SpatialAq) -> Self

Spatial AQ makes adjustments within each frame based on spatial variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if spatial AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply spatial AQ using the specified strength. Disabled: MediaLive won’t apply spatial AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn’t apply spatial AQ.

pub fn set_spatial_aq(self, input: Option<H264SpatialAq>) -> Self

pub fn subgop_length(self, input: H264SubGopLength) -> Self

If set to fixed, use gopNumBFrames B-frames per sub-GOP. If set to dynamic, optimize the number of B-frames used for each sub-GOP to improve visual quality.

pub fn set_subgop_length(self, input: Option<H264SubGopLength>) -> Self

pub fn syntax(self, input: H264Syntax) -> Self

Produces a bitstream compliant with SMPTE RP-2027.

pub fn set_syntax(self, input: Option<H264Syntax>) -> Self

pub fn temporal_aq(self, input: H264TemporalAq) -> Self

Temporal makes adjustments within each frame based on temporal variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if temporal AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply temporal AQ using the specified strength. Disabled: MediaLive won’t apply temporal AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn’t apply temporal AQ.

pub fn set_temporal_aq(self, input: Option<H264TemporalAq>) -> Self

pub fn timecode_insertion(self, input: H264TimecodeInsertionBehavior) -> Self

Determines how timecodes should be inserted into the video elementary stream.

‘disabled’: Do not include timecodes

‘picTimingSei’: Pass through picture timing SEI messages from the source specified in Timecode Config

pub fn set_timecode_insertion(
self,
input: Option<H264TimecodeInsertionBehavior>
) -> Self

pub fn build(self) -> H264Settings

Consumes the builder and constructs a H264Settings