Struct aws_sdk_medialive::model::h265_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for H265Settings

Implementations

impl Builder

pub fn adaptive_quantization(self, input: H265AdaptiveQuantization) -> Self

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

pub fn set_adaptive_quantization(
self,
input: Option<H265AdaptiveQuantization>
) -> Self

pub fn afd_signaling(self, input: AfdSignaling) -> Self

Indicates that AFD values will be written into the output stream. If afdSignaling is “auto”, the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to “fixed”, the AFD value will be the value configured in the fixedAfd parameter.

pub fn set_afd_signaling(self, input: Option<AfdSignaling>) -> Self

pub fn alternative_transfer_function(
self,
input: H265AlternativeTransferFunction
) -> Self

Whether or not EML should insert an Alternative Transfer Function SEI message to support backwards compatibility with non-HDR decoders and displays.

pub fn set_alternative_transfer_function(
self,
input: Option<H265AlternativeTransferFunction>
) -> Self

pub fn bitrate(self, input: i32) -> Self

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

pub fn set_bitrate(self, input: Option<i32>) -> Self

pub fn buf_size(self, input: i32) -> Self

Size of buffer (HRD buffer model) in bits.

pub fn set_buf_size(self, input: Option<i32>) -> Self

pub fn color_metadata(self, input: H265ColorMetadata) -> Self

Includes colorspace metadata in the output.

pub fn set_color_metadata(self, input: Option<H265ColorMetadata>) -> Self

pub fn color_space_settings(self, input: H265ColorSpaceSettings) -> Self

Color Space settings

pub fn set_color_space_settings(
self,
input: Option<H265ColorSpaceSettings>
) -> Self

pub fn filter_settings(self, input: H265FilterSettings) -> Self

Optional filters that you can apply to an encode.

pub fn set_filter_settings(self, input: Option<H265FilterSettings>) -> Self

pub fn fixed_afd(self, input: FixedAfd) -> Self

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to ‘Fixed’.

pub fn set_fixed_afd(self, input: Option<FixedAfd>) -> Self

pub fn flicker_aq(self, input: H265FlickerAq) -> Self

If set to enabled, adjust quantization within each frame to reduce flicker or ‘pop’ on I-frames.

pub fn set_flicker_aq(self, input: Option<H265FlickerAq>) -> Self

pub fn framerate_denominator(self, input: i32) -> Self

Framerate denominator.

pub fn set_framerate_denominator(self, input: Option<i32>) -> Self

pub fn framerate_numerator(self, input: i32) -> Self

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

pub fn set_framerate_numerator(self, input: Option<i32>) -> Self

pub fn gop_closed_cadence(self, input: i32) -> Self

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

pub fn set_gop_closed_cadence(self, input: Option<i32>) -> Self

pub fn gop_size(self, input: f64) -> Self

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

pub fn set_gop_size(self, input: Option<f64>) -> Self

pub fn gop_size_units(self, input: H265GopSizeUnits) -> Self

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

pub fn set_gop_size_units(self, input: Option<H265GopSizeUnits>) -> Self

pub fn level(self, input: H265Level) -> Self

H.265 Level.

pub fn set_level(self, input: Option<H265Level>) -> Self

pub fn look_ahead_rate_control(self, input: H265LookAheadRateControl) -> Self

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

pub fn set_look_ahead_rate_control(
self,
input: Option<H265LookAheadRateControl>
) -> Self

pub fn max_bitrate(self, input: i32) -> Self

For QVBR: See the tooltip for Quality level

pub fn set_max_bitrate(self, input: Option<i32>) -> Self

pub fn min_i_interval(self, input: i32) -> Self

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

pub fn set_min_i_interval(self, input: Option<i32>) -> Self

pub fn par_denominator(self, input: i32) -> Self

Pixel Aspect Ratio denominator.

pub fn set_par_denominator(self, input: Option<i32>) -> Self

pub fn par_numerator(self, input: i32) -> Self

Pixel Aspect Ratio numerator.

pub fn set_par_numerator(self, input: Option<i32>) -> Self

pub fn profile(self, input: H265Profile) -> Self

H.265 Profile.

pub fn set_profile(self, input: Option<H265Profile>) -> Self

pub fn qvbr_quality_level(self, input: i32) -> Self

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. Set values for the QVBR quality level field and Max bitrate field that suit your most important viewing devices. Recommended values are:

Primary screen: Quality level: 8 to 10. Max bitrate: 4M

PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M

Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M

pub fn set_qvbr_quality_level(self, input: Option<i32>) -> Self

pub fn rate_control_mode(self, input: H265RateControlMode) -> Self

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

pub fn set_rate_control_mode(self, input: Option<H265RateControlMode>) -> Self

pub fn scan_type(self, input: H265ScanType) -> Self

Sets the scan type of the output to progressive or top-field-first interlaced.

pub fn set_scan_type(self, input: Option<H265ScanType>) -> Self

pub fn scene_change_detect(self, input: H265SceneChangeDetect) -> Self

Scene change detection.

pub fn set_scene_change_detect(
self,
input: Option<H265SceneChangeDetect>
) -> Self

pub fn slices(self, input: i32) -> Self

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

pub fn set_slices(self, input: Option<i32>) -> Self

pub fn tier(self, input: H265Tier) -> Self

H.265 Tier.

pub fn set_tier(self, input: Option<H265Tier>) -> Self

pub fn timecode_insertion(self, input: H265TimecodeInsertionBehavior) -> Self

Determines how timecodes should be inserted into the video elementary stream.

‘disabled’: Do not include timecodes

‘picTimingSei’: Pass through picture timing SEI messages from the source specified in Timecode Config

pub fn set_timecode_insertion(
self,
input: Option<H265TimecodeInsertionBehavior>
) -> Self

pub fn build(self) -> H265Settings

Consumes the builder and constructs a H265Settings