Struct aws_sdk_medialive::model::hls_group_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for HlsGroupSettings

Implementations

impl Builder

pub fn ad_markers(self, input: impl Into<HlsAdMarkers>) -> Self

pub fn set_ad_markers(self, input: Option<Vec<HlsAdMarkers>>) -> Self

pub fn base_url_content(self, input: impl Into<String>) -> Self

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

pub fn set_base_url_content(self, input: Option<String>) -> Self

pub fn base_url_content1(self, input: impl Into<String>) -> Self

Optional. One value per output group. This field is required only if you are completing Base URL content A, and the downstream system has notified you that the media files for pipeline 1 of all outputs are in a location different from the media files for pipeline 0.

pub fn set_base_url_content1(self, input: Option<String>) -> Self

pub fn base_url_manifest(self, input: impl Into<String>) -> Self

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

pub fn set_base_url_manifest(self, input: Option<String>) -> Self

pub fn base_url_manifest1(self, input: impl Into<String>) -> Self

Optional. One value per output group. Complete this field only if you are completing Base URL manifest A, and the downstream system has notified you that the child manifest files for pipeline 1 of all outputs are in a location different from the child manifest files for pipeline 0.

pub fn set_base_url_manifest1(self, input: Option<String>) -> Self

pub fn caption_language_mappings(
self,
input: impl Into<CaptionLanguageMapping>
) -> Self

pub fn set_caption_language_mappings(
self,
input: Option<Vec<CaptionLanguageMapping>>
) -> Self

pub fn caption_language_setting(self, input: HlsCaptionLanguageSetting) -> Self

Applies only to 608 Embedded output captions. insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. none: Include CLOSED-CAPTIONS=NONE line in the manifest. omit: Omit any CLOSED-CAPTIONS line from the manifest.

pub fn set_caption_language_setting(
self,
input: Option<HlsCaptionLanguageSetting>
) -> Self

pub fn client_cache(self, input: HlsClientCache) -> Self

When set to “disabled”, sets the #EXT-X-ALLOW-CACHE:no tag in the manifest, which prevents clients from saving media segments for later replay.

pub fn set_client_cache(self, input: Option<HlsClientCache>) -> Self

pub fn codec_specification(self, input: HlsCodecSpecification) -> Self

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

pub fn set_codec_specification(
self,
input: Option<HlsCodecSpecification>
) -> Self

pub fn constant_iv(self, input: impl Into<String>) -> Self

For use with encryptionType. This is a 128-bit, 16-byte hex value represented by a 32-character text string. If ivSource is set to “explicit” then this parameter is required and is used as the IV for encryption.

pub fn set_constant_iv(self, input: Option<String>) -> Self

pub fn destination(self, input: OutputLocationRef) -> Self

A directory or HTTP destination for the HLS segments, manifest files, and encryption keys (if enabled).

pub fn set_destination(self, input: Option<OutputLocationRef>) -> Self

pub fn directory_structure(self, input: HlsDirectoryStructure) -> Self

Place segments in subdirectories.

pub fn set_directory_structure(
self,
input: Option<HlsDirectoryStructure>
) -> Self

pub fn discontinuity_tags(self, input: HlsDiscontinuityTags) -> Self

Specifies whether to insert EXT-X-DISCONTINUITY tags in the HLS child manifests for this output group. Typically, choose Insert because these tags are required in the manifest (according to the HLS specification) and serve an important purpose. Choose Never Insert only if the downstream system is doing real-time failover (without using the MediaLive automatic failover feature) and only if that downstream system has advised you to exclude the tags.

pub fn set_discontinuity_tags(self, input: Option<HlsDiscontinuityTags>) -> Self

pub fn encryption_type(self, input: HlsEncryptionType) -> Self

Encrypts the segments with the given encryption scheme. Exclude this parameter if no encryption is desired.

pub fn set_encryption_type(self, input: Option<HlsEncryptionType>) -> Self

pub fn hls_cdn_settings(self, input: HlsCdnSettings) -> Self

Parameters that control interactions with the CDN.

pub fn set_hls_cdn_settings(self, input: Option<HlsCdnSettings>) -> Self

pub fn hls_id3_segment_tagging(self, input: HlsId3SegmentTaggingState) -> Self

State of HLS ID3 Segment Tagging

pub fn set_hls_id3_segment_tagging(
self,
input: Option<HlsId3SegmentTaggingState>
) -> Self

pub fn i_frame_only_playlists(self, input: IFrameOnlyPlaylistType) -> Self

DISABLED: Do not create an I-frame-only manifest, but do create the master and media manifests (according to the Output Selection field). STANDARD: Create an I-frame-only manifest for each output that contains video, as well as the other manifests (according to the Output Selection field). The I-frame manifest contains a #EXT-X-I-FRAMES-ONLY tag to indicate it is I-frame only, and one or more #EXT-X-BYTERANGE entries identifying the I-frame position. For example, #EXT-X-BYTERANGE:160364@1461888“

pub fn set_i_frame_only_playlists(
self,
input: Option<IFrameOnlyPlaylistType>
) -> Self

pub fn incomplete_segment_behavior(
self,
input: HlsIncompleteSegmentBehavior
) -> Self

Specifies whether to include the final (incomplete) segment in the media output when the pipeline stops producing output because of a channel stop, a channel pause or a loss of input to the pipeline. Auto means that MediaLive decides whether to include the final segment, depending on the channel class and the types of output groups. Suppress means to never include the incomplete segment. We recommend you choose Auto and let MediaLive control the behavior.

pub fn set_incomplete_segment_behavior(
self,
input: Option<HlsIncompleteSegmentBehavior>
) -> Self

pub fn index_n_segments(self, input: i32) -> Self

Applies only if Mode field is LIVE. Specifies the maximum number of segments in the media manifest file. After this maximum, older segments are removed from the media manifest. This number must be smaller than the number in the Keep Segments field.

pub fn set_index_n_segments(self, input: Option<i32>) -> Self

pub fn input_loss_action(self, input: InputLossActionForHlsOut) -> Self

Parameter that control output group behavior on input loss.

pub fn set_input_loss_action(
self,
input: Option<InputLossActionForHlsOut>
) -> Self

pub fn iv_in_manifest(self, input: HlsIvInManifest) -> Self

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If set to “include”, IV is listed in the manifest, otherwise the IV is not in the manifest.

pub fn set_iv_in_manifest(self, input: Option<HlsIvInManifest>) -> Self

pub fn iv_source(self, input: HlsIvSource) -> Self

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If this setting is “followsSegmentNumber”, it will cause the IV to change every segment (to match the segment number). If this is set to “explicit”, you must enter a constantIv value.

pub fn set_iv_source(self, input: Option<HlsIvSource>) -> Self

pub fn keep_segments(self, input: i32) -> Self

Applies only if Mode field is LIVE. Specifies the number of media segments to retain in the destination directory. This number should be bigger than indexNSegments (Num segments). We recommend (value = (2 x indexNsegments) + 1). If this “keep segments” number is too low, the following might happen: the player is still reading a media manifest file that lists this segment, but that segment has been removed from the destination directory (as directed by indexNSegments). This situation would result in a 404 HTTP error on the player.

pub fn set_keep_segments(self, input: Option<i32>) -> Self

pub fn key_format(self, input: impl Into<String>) -> Self

The value specifies how the key is represented in the resource identified by the URI. If parameter is absent, an implicit value of “identity” is used. A reverse DNS string can also be given.

pub fn set_key_format(self, input: Option<String>) -> Self

pub fn key_format_versions(self, input: impl Into<String>) -> Self

Either a single positive integer version value or a slash delimited list of version values (1/2/3).

pub fn set_key_format_versions(self, input: Option<String>) -> Self

pub fn key_provider_settings(self, input: KeyProviderSettings) -> Self

The key provider settings.

pub fn set_key_provider_settings(
self,
input: Option<KeyProviderSettings>
) -> Self

pub fn manifest_compression(self, input: HlsManifestCompression) -> Self

When set to gzip, compresses HLS playlist.

pub fn set_manifest_compression(
self,
input: Option<HlsManifestCompression>
) -> Self

pub fn manifest_duration_format(self, input: HlsManifestDurationFormat) -> Self

Indicates whether the output manifest should use floating point or integer values for segment duration.

pub fn set_manifest_duration_format(
self,
input: Option<HlsManifestDurationFormat>
) -> Self

pub fn min_segment_length(self, input: i32) -> Self

When set, minimumSegmentLength is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

pub fn set_min_segment_length(self, input: Option<i32>) -> Self

pub fn mode(self, input: HlsMode) -> Self

If “vod”, all segments are indexed and kept permanently in the destination and manifest. If “live”, only the number segments specified in keepSegments and indexNSegments are kept; newer segments replace older segments, which may prevent players from rewinding all the way to the beginning of the event. VOD mode uses HLS EXT-X-PLAYLIST-TYPE of EVENT while the channel is running, converting it to a “VOD” type manifest on completion of the stream.

pub fn set_mode(self, input: Option<HlsMode>) -> Self

pub fn output_selection(self, input: HlsOutputSelection) -> Self

MANIFESTS_AND_SEGMENTS: Generates manifests (master manifest, if applicable, and media manifests) for this output group. VARIANT_MANIFESTS_AND_SEGMENTS: Generates media manifests for this output group, but not a master manifest. SEGMENTS_ONLY: Does not generate any manifests for this output group.

pub fn set_output_selection(self, input: Option<HlsOutputSelection>) -> Self

pub fn program_date_time(self, input: HlsProgramDateTime) -> Self

Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestampOffset.

pub fn set_program_date_time(self, input: Option<HlsProgramDateTime>) -> Self

pub fn program_date_time_period(self, input: i32) -> Self

Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

pub fn set_program_date_time_period(self, input: Option<i32>) -> Self

pub fn redundant_manifest(self, input: HlsRedundantManifest) -> Self

ENABLED: The master manifest (.m3u8 file) for each pipeline includes information about both pipelines: first its own media files, then the media files of the other pipeline. This feature allows playout device that support stale manifest detection to switch from one manifest to the other, when the current manifest seems to be stale. There are still two destinations and two master manifests, but both master manifests reference the media files from both pipelines. DISABLED: The master manifest (.m3u8 file) for each pipeline includes information about its own pipeline only. For an HLS output group with MediaPackage as the destination, the DISABLED behavior is always followed. MediaPackage regenerates the manifests it serves to players so a redundant manifest from MediaLive is irrelevant.

pub fn set_redundant_manifest(self, input: Option<HlsRedundantManifest>) -> Self

pub fn segment_length(self, input: i32) -> Self

Length of MPEG-2 Transport Stream segments to create (in seconds). Note that segments will end on the next keyframe after this number of seconds, so actual segment length may be longer.

pub fn set_segment_length(self, input: Option<i32>) -> Self

pub fn segmentation_mode(self, input: HlsSegmentationMode) -> Self

useInputSegmentation has been deprecated. The configured segment size is always used.

pub fn set_segmentation_mode(self, input: Option<HlsSegmentationMode>) -> Self

pub fn segments_per_subdirectory(self, input: i32) -> Self

Number of segments to write to a subdirectory before starting a new one. directoryStructure must be subdirectoryPerStream for this setting to have an effect.

pub fn set_segments_per_subdirectory(self, input: Option<i32>) -> Self

pub fn stream_inf_resolution(self, input: HlsStreamInfResolution) -> Self

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

pub fn set_stream_inf_resolution(
self,
input: Option<HlsStreamInfResolution>
) -> Self

pub fn timed_metadata_id3_frame(self, input: HlsTimedMetadataId3Frame) -> Self

Indicates ID3 frame that has the timecode.

pub fn set_timed_metadata_id3_frame(
self,
input: Option<HlsTimedMetadataId3Frame>
) -> Self

pub fn timed_metadata_id3_period(self, input: i32) -> Self

Timed Metadata interval in seconds.

pub fn set_timed_metadata_id3_period(self, input: Option<i32>) -> Self

pub fn timestamp_delta_milliseconds(self, input: i32) -> Self

Provides an extra millisecond delta offset to fine tune the timestamps.

pub fn set_timestamp_delta_milliseconds(self, input: Option<i32>) -> Self

pub fn ts_file_mode(self, input: HlsTsFileMode) -> Self

SEGMENTED_FILES: Emit the program as segments - multiple .ts media files. SINGLE_FILE: Applies only if Mode field is VOD. Emit the program as a single .ts media file. The media manifest includes #EXT-X-BYTERANGE tags to index segments for playback. A typical use for this value is when sending the output to AWS Elemental MediaConvert, which can accept only a single media file. Playback while the channel is running is not guaranteed due to HTTP server caching.

pub fn set_ts_file_mode(self, input: Option<HlsTsFileMode>) -> Self

pub fn build(self) -> HlsGroupSettings

Consumes the builder and constructs a HlsGroupSettings