Struct aws_sdk_medialive::model::m2ts_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for M2tsSettings

Implementations

impl Builder

pub fn absent_input_audio_behavior(
self,
input: M2tsAbsentInputAudioBehavior
) -> Self

When set to drop, output audio streams will be removed from the program if the selected input audio stream is removed from the input. This allows the output audio configuration to dynamically change based on input configuration. If this is set to encodeSilence, all output audio streams will output encoded silence when not connected to an active input stream.

pub fn set_absent_input_audio_behavior(
self,
input: Option<M2tsAbsentInputAudioBehavior>
) -> Self

pub fn arib(self, input: M2tsArib) -> Self

When set to enabled, uses ARIB-compliant field muxing and removes video descriptor.

pub fn set_arib(self, input: Option<M2tsArib>) -> Self

pub fn arib_captions_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for ARIB Captions in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_arib_captions_pid(self, input: Option<String>) -> Self

pub fn arib_captions_pid_control(
self,
input: M2tsAribCaptionsPidControl
) -> Self

If set to auto, pid number used for ARIB Captions will be auto-selected from unused pids. If set to useConfigured, ARIB Captions will be on the configured pid number.

pub fn set_arib_captions_pid_control(
self,
input: Option<M2tsAribCaptionsPidControl>
) -> Self

pub fn audio_buffer_model(self, input: M2tsAudioBufferModel) -> Self

When set to dvb, uses DVB buffer model for Dolby Digital audio. When set to atsc, the ATSC model is used.

pub fn set_audio_buffer_model(self, input: Option<M2tsAudioBufferModel>) -> Self

pub fn audio_frames_per_pes(self, input: i32) -> Self

The number of audio frames to insert for each PES packet.

pub fn set_audio_frames_per_pes(self, input: Option<i32>) -> Self

pub fn audio_pids(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_audio_pids(self, input: Option<String>) -> Self

pub fn audio_stream_type(self, input: M2tsAudioStreamType) -> Self

When set to atsc, uses stream type = 0x81 for AC3 and stream type = 0x87 for EAC3. When set to dvb, uses stream type = 0x06.

pub fn set_audio_stream_type(self, input: Option<M2tsAudioStreamType>) -> Self

pub fn bitrate(self, input: i32) -> Self

The output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate.

pub fn set_bitrate(self, input: Option<i32>) -> Self

pub fn buffer_model(self, input: M2tsBufferModel) -> Self

Controls the timing accuracy for output network traffic. Leave as MULTIPLEX to ensure accurate network packet timing. Or set to NONE, which might result in lower latency but will result in more variability in output network packet timing. This variability might cause interruptions, jitter, or bursty behavior in your playback or receiving devices.

pub fn set_buffer_model(self, input: Option<M2tsBufferModel>) -> Self

pub fn cc_descriptor(self, input: M2tsCcDescriptor) -> Self

When set to enabled, generates captionServiceDescriptor in PMT.

pub fn set_cc_descriptor(self, input: Option<M2tsCcDescriptor>) -> Self

pub fn dvb_nit_settings(self, input: DvbNitSettings) -> Self

Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

pub fn set_dvb_nit_settings(self, input: Option<DvbNitSettings>) -> Self

pub fn dvb_sdt_settings(self, input: DvbSdtSettings) -> Self

Inserts DVB Service Description Table (SDT) at the specified table repetition interval.

pub fn set_dvb_sdt_settings(self, input: Option<DvbSdtSettings>) -> Self

pub fn dvb_sub_pids(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source DVB Subtitle data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_dvb_sub_pids(self, input: Option<String>) -> Self

pub fn dvb_tdt_settings(self, input: DvbTdtSettings) -> Self

Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

pub fn set_dvb_tdt_settings(self, input: Option<DvbTdtSettings>) -> Self

pub fn dvb_teletext_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source DVB Teletext data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_dvb_teletext_pid(self, input: Option<String>) -> Self

pub fn ebif(self, input: M2tsEbifControl) -> Self

If set to passthrough, passes any EBIF data from the input source to this output.

pub fn set_ebif(self, input: Option<M2tsEbifControl>) -> Self

pub fn ebp_audio_interval(self, input: M2tsAudioInterval) -> Self

When videoAndFixedIntervals is selected, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. Only available when EBP Cablelabs segmentation markers are selected. Partitions 1 and 2 will always follow the video interval.

pub fn set_ebp_audio_interval(self, input: Option<M2tsAudioInterval>) -> Self

pub fn ebp_lookahead_ms(self, input: i32) -> Self

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is “stretched” to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

pub fn set_ebp_lookahead_ms(self, input: Option<i32>) -> Self

pub fn ebp_placement(self, input: M2tsEbpPlacement) -> Self

Controls placement of EBP on Audio PIDs. If set to videoAndAudioPids, EBP markers will be placed on the video PID and all audio PIDs. If set to videoPid, EBP markers will be placed on only the video PID.

pub fn set_ebp_placement(self, input: Option<M2tsEbpPlacement>) -> Self

pub fn ecm_pid(self, input: impl Into<String>) -> Self

This field is unused and deprecated.

pub fn set_ecm_pid(self, input: Option<String>) -> Self

pub fn es_rate_in_pes(self, input: M2tsEsRateInPes) -> Self

Include or exclude the ES Rate field in the PES header.

pub fn set_es_rate_in_pes(self, input: Option<M2tsEsRateInPes>) -> Self

pub fn etv_platform_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source ETV Platform data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_etv_platform_pid(self, input: Option<String>) -> Self

pub fn etv_signal_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source ETV Signal data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_etv_signal_pid(self, input: Option<String>) -> Self

pub fn fragment_time(self, input: f64) -> Self

The length in seconds of each fragment. Only used with EBP markers.

pub fn set_fragment_time(self, input: Option<f64>) -> Self

pub fn klv(self, input: M2tsKlv) -> Self

If set to passthrough, passes any KLV data from the input source to this output.

pub fn set_klv(self, input: Option<M2tsKlv>) -> Self

pub fn klv_data_pids(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source KLV data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_klv_data_pids(self, input: Option<String>) -> Self

pub fn nielsen_id3_behavior(self, input: M2tsNielsenId3Behavior) -> Self

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

pub fn set_nielsen_id3_behavior(
self,
input: Option<M2tsNielsenId3Behavior>
) -> Self

pub fn null_packet_bitrate(self, input: f64) -> Self

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

pub fn set_null_packet_bitrate(self, input: Option<f64>) -> Self

pub fn pat_interval(self, input: i32) -> Self

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

pub fn set_pat_interval(self, input: Option<i32>) -> Self

pub fn pcr_control(self, input: M2tsPcrControl) -> Self

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

pub fn set_pcr_control(self, input: Option<M2tsPcrControl>) -> Self

pub fn pcr_period(self, input: i32) -> Self

Maximum time in milliseconds between Program Clock Reference (PCRs) inserted into the transport stream.

pub fn set_pcr_period(self, input: Option<i32>) -> Self

pub fn pcr_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_pcr_pid(self, input: Option<String>) -> Self

pub fn pmt_interval(self, input: i32) -> Self

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

pub fn set_pmt_interval(self, input: Option<i32>) -> Self

pub fn pmt_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_pmt_pid(self, input: Option<String>) -> Self

pub fn program_num(self, input: i32) -> Self

The value of the program number field in the Program Map Table.

pub fn set_program_num(self, input: Option<i32>) -> Self

pub fn rate_mode(self, input: M2tsRateMode) -> Self

When vbr, does not insert null packets into transport stream to fill specified bitrate. The bitrate setting acts as the maximum bitrate when vbr is set.

pub fn set_rate_mode(self, input: Option<M2tsRateMode>) -> Self

pub fn scte27_pids(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for input source SCTE-27 data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_scte27_pids(self, input: Option<String>) -> Self

pub fn scte35_control(self, input: M2tsScte35Control) -> Self

Optionally pass SCTE-35 signals from the input source to this output.

pub fn set_scte35_control(self, input: Option<M2tsScte35Control>) -> Self

pub fn scte35_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_scte35_pid(self, input: Option<String>) -> Self

pub fn segmentation_markers(self, input: M2tsSegmentationMarkers) -> Self

Inserts segmentation markers at each segmentationTime period. raiSegstart sets the Random Access Indicator bit in the adaptation field. raiAdapt sets the RAI bit and adds the current timecode in the private data bytes. psiSegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebpLegacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

pub fn set_segmentation_markers(
self,
input: Option<M2tsSegmentationMarkers>
) -> Self

pub fn segmentation_style(self, input: M2tsSegmentationStyle) -> Self

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of “resetCadence” is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of $segmentationTime seconds. When a segmentation style of “maintainCadence” is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationTime seconds. Note that EBP lookahead is a slight exception to this rule.

pub fn set_segmentation_style(
self,
input: Option<M2tsSegmentationStyle>
) -> Self

pub fn segmentation_time(self, input: f64) -> Self

The length in seconds of each segment. Required unless markers is set to none.

pub fn set_segmentation_time(self, input: Option<f64>) -> Self

pub fn timed_metadata_behavior(self, input: M2tsTimedMetadataBehavior) -> Self

When set to passthrough, timed metadata will be passed through from input to output.

pub fn set_timed_metadata_behavior(
self,
input: Option<M2tsTimedMetadataBehavior>
) -> Self

pub fn timed_metadata_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_timed_metadata_pid(self, input: Option<String>) -> Self

pub fn transport_stream_id(self, input: i32) -> Self

The value of the transport stream ID field in the Program Map Table.

pub fn set_transport_stream_id(self, input: Option<i32>) -> Self

pub fn video_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_video_pid(self, input: Option<String>) -> Self

pub fn build(self) -> M2tsSettings

Consumes the builder and constructs a M2tsSettings