Struct aws_sdk_medialive::model::m3u8_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for M3u8Settings

Implementations

impl Builder

pub fn audio_frames_per_pes(self, input: i32) -> Self

The number of audio frames to insert for each PES packet.

pub fn set_audio_frames_per_pes(self, input: Option<i32>) -> Self

pub fn audio_pids(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values.

pub fn set_audio_pids(self, input: Option<String>) -> Self

pub fn ecm_pid(self, input: impl Into<String>) -> Self

This parameter is unused and deprecated.

pub fn set_ecm_pid(self, input: Option<String>) -> Self

pub fn nielsen_id3_behavior(self, input: M3u8NielsenId3Behavior) -> Self

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

pub fn set_nielsen_id3_behavior(
self,
input: Option<M3u8NielsenId3Behavior>
) -> Self

pub fn pat_interval(self, input: i32) -> Self

The number of milliseconds between instances of this table in the output transport stream. A value of "0" writes out the PMT once per segment file.

pub fn set_pat_interval(self, input: Option<i32>) -> Self

pub fn pcr_control(self, input: M3u8PcrControl) -> Self

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

pub fn set_pcr_control(self, input: Option<M3u8PcrControl>) -> Self

pub fn pcr_period(self, input: i32) -> Self

Maximum time in milliseconds between Program Clock References (PCRs) inserted into the transport stream.

pub fn set_pcr_period(self, input: Option<i32>) -> Self

pub fn pcr_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value.

pub fn set_pcr_pid(self, input: Option<String>) -> Self

pub fn pmt_interval(self, input: i32) -> Self

The number of milliseconds between instances of this table in the output transport stream. A value of "0" writes out the PMT once per segment file.

pub fn set_pmt_interval(self, input: Option<i32>) -> Self

pub fn pmt_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value.

pub fn set_pmt_pid(self, input: Option<String>) -> Self

pub fn program_num(self, input: i32) -> Self

The value of the program number field in the Program Map Table.

pub fn set_program_num(self, input: Option<i32>) -> Self

pub fn scte35_behavior(self, input: M3u8Scte35Behavior) -> Self

If set to passthrough, passes any SCTE-35 signals from the input source to this output.

pub fn set_scte35_behavior(self, input: Option<M3u8Scte35Behavior>) -> Self

pub fn scte35_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value.

pub fn set_scte35_pid(self, input: Option<String>) -> Self

pub fn timed_metadata_behavior(self, input: M3u8TimedMetadataBehavior) -> Self

When set to passthrough, timed metadata is passed through from input to output.

pub fn set_timed_metadata_behavior(
self,
input: Option<M3u8TimedMetadataBehavior>
) -> Self

pub fn timed_metadata_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pub fn set_timed_metadata_pid(self, input: Option<String>) -> Self

pub fn transport_stream_id(self, input: i32) -> Self

The value of the transport stream ID field in the Program Map Table.

pub fn set_transport_stream_id(self, input: Option<i32>) -> Self

pub fn video_pid(self, input: impl Into<String>) -> Self

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value.

pub fn set_video_pid(self, input: Option<String>) -> Self

pub fn build(self) -> M3u8Settings

Consumes the builder and constructs a M3u8Settings