Struct aws_sdk_medialive::model::mpeg2_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Mpeg2Settings

Implementations

impl Builder

pub fn adaptive_quantization(self, input: Mpeg2AdaptiveQuantization) -> Self

Choose Off to disable adaptive quantization. Or choose another value to enable the quantizer and set its strength. The strengths are: Auto, Off, Low, Medium, High. When you enable this field, MediaLive allows intra-frame quantizers to vary, which might improve visual quality.

pub fn set_adaptive_quantization(
self,
input: Option<Mpeg2AdaptiveQuantization>
) -> Self

pub fn afd_signaling(self, input: AfdSignaling) -> Self

Indicates the AFD values that MediaLive will write into the video encode. If you do not know what AFD signaling is, or if your downstream system has not given you guidance, choose AUTO. AUTO: MediaLive will try to preserve the input AFD value (in cases where multiple AFD values are valid). FIXED: MediaLive will use the value you specify in fixedAFD.

pub fn set_afd_signaling(self, input: Option<AfdSignaling>) -> Self

pub fn color_metadata(self, input: Mpeg2ColorMetadata) -> Self

Specifies whether to include the color space metadata. The metadata describes the color space that applies to the video (the colorSpace field). We recommend that you insert the metadata.

pub fn set_color_metadata(self, input: Option<Mpeg2ColorMetadata>) -> Self

pub fn color_space(self, input: Mpeg2ColorSpace) -> Self

Choose the type of color space conversion to apply to the output. For detailed information on setting up both the input and the output to obtain the desired color space in the output, see the section on "MediaLive Features - Video - color space" in the MediaLive User Guide. PASSTHROUGH: Keep the color space of the input content - do not convert it. AUTO:Convert all content that is SD to rec 601, and convert all content that is HD to rec 709.

pub fn set_color_space(self, input: Option<Mpeg2ColorSpace>) -> Self

pub fn display_aspect_ratio(self, input: Mpeg2DisplayRatio) -> Self

Sets the pixel aspect ratio for the encode.

pub fn set_display_aspect_ratio(self, input: Option<Mpeg2DisplayRatio>) -> Self

pub fn filter_settings(self, input: Mpeg2FilterSettings) -> Self

Optionally specify a noise reduction filter, which can improve quality of compressed content. If you do not choose a filter, no filter will be applied. TEMPORAL: This filter is useful for both source content that is noisy (when it has excessive digital artifacts) and source content that is clean. When the content is noisy, the filter cleans up the source content before the encoding phase, with these two effects: First, it improves the output video quality because the content has been cleaned up. Secondly, it decreases the bandwidth because MediaLive does not waste bits on encoding noise. When the content is reasonably clean, the filter tends to decrease the bitrate.

pub fn set_filter_settings(self, input: Option<Mpeg2FilterSettings>) -> Self

pub fn fixed_afd(self, input: FixedAfd) -> Self

Complete this field only when afdSignaling is set to FIXED. Enter the AFD value (4 bits) to write on all frames of the video encode.

pub fn set_fixed_afd(self, input: Option<FixedAfd>) -> Self

pub fn framerate_denominator(self, input: i32) -> Self

description“: “The framerate denominator. For example, 1001. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

pub fn set_framerate_denominator(self, input: Option<i32>) -> Self

pub fn framerate_numerator(self, input: i32) -> Self

The framerate numerator. For example, 24000. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

pub fn set_framerate_numerator(self, input: Option<i32>) -> Self

pub fn gop_closed_cadence(self, input: i32) -> Self

MPEG2: default is open GOP.

pub fn set_gop_closed_cadence(self, input: Option<i32>) -> Self

pub fn gop_num_b_frames(self, input: i32) -> Self

Relates to the GOP structure. The number of B-frames between reference frames. If you do not know what a B-frame is, use the default.

pub fn set_gop_num_b_frames(self, input: Option<i32>) -> Self

pub fn gop_size(self, input: f64) -> Self

Relates to the GOP structure. The GOP size (keyframe interval) in the units specified in gopSizeUnits. If you do not know what GOP is, use the default. If gopSizeUnits is frames, then the gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, the gopSize must be greater than 0, but does not need to be an integer.

pub fn set_gop_size(self, input: Option<f64>) -> Self

pub fn gop_size_units(self, input: Mpeg2GopSizeUnits) -> Self

Relates to the GOP structure. Specifies whether the gopSize is specified in frames or seconds. If you do not plan to change the default gopSize, leave the default. If you specify SECONDS, MediaLive will internally convert the gop size to a frame count.

pub fn set_gop_size_units(self, input: Option<Mpeg2GopSizeUnits>) -> Self

pub fn scan_type(self, input: Mpeg2ScanType) -> Self

Set the scan type of the output to PROGRESSIVE or INTERLACED (top field first).

pub fn set_scan_type(self, input: Option<Mpeg2ScanType>) -> Self

pub fn subgop_length(self, input: Mpeg2SubGopLength) -> Self

Relates to the GOP structure. If you do not know what GOP is, use the default. FIXED: Set the number of B-frames in each sub-GOP to the value in gopNumBFrames. DYNAMIC: Let MediaLive optimize the number of B-frames in each sub-GOP, to improve visual quality.

pub fn set_subgop_length(self, input: Option<Mpeg2SubGopLength>) -> Self

pub fn timecode_insertion(self, input: Mpeg2TimecodeInsertionBehavior) -> Self

Determines how MediaLive inserts timecodes in the output video. For detailed information about setting up the input and the output for a timecode, see the section on "MediaLive Features - Timecode configuration" in the MediaLive User Guide. DISABLED: do not include timecodes. GOP_TIMECODE: Include timecode metadata in the GOP header.

pub fn set_timecode_insertion(
self,
input: Option<Mpeg2TimecodeInsertionBehavior>
) -> Self

pub fn build(self) -> Mpeg2Settings

Consumes the builder and constructs a Mpeg2Settings