Struct aws_sdk_medialive::model::ms_smooth_group_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MsSmoothGroupSettings

Implementations

impl Builder

pub fn acquisition_point_id(self, input: impl Into<String>) -> Self

The ID to include in each message in the sparse track. Ignored if sparseTrackType is NONE.

pub fn set_acquisition_point_id(self, input: Option<String>) -> Self

pub fn audio_only_timecode_control(
self,
input: SmoothGroupAudioOnlyTimecodeControl
) -> Self

If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

pub fn set_audio_only_timecode_control(
self,
input: Option<SmoothGroupAudioOnlyTimecodeControl>
) -> Self

pub fn certificate_mode(self, input: SmoothGroupCertificateMode) -> Self

If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

pub fn set_certificate_mode(
self,
input: Option<SmoothGroupCertificateMode>
) -> Self

pub fn connection_retry_interval(self, input: i32) -> Self

Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

pub fn set_connection_retry_interval(self, input: Option<i32>) -> Self

pub fn destination(self, input: OutputLocationRef) -> Self

Smooth Streaming publish point on an IIS server. Elemental Live acts as a “Push” encoder to IIS.

pub fn set_destination(self, input: Option<OutputLocationRef>) -> Self

pub fn event_id(self, input: impl Into<String>) -> Self

MS Smooth event ID to be sent to the IIS server. Should only be specified if eventIdMode is set to useConfigured.

pub fn set_event_id(self, input: Option<String>) -> Self

pub fn event_id_mode(self, input: SmoothGroupEventIdMode) -> Self

Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run. Options:

“useConfigured” - use the value provided in eventId

“useTimestamp” - generate and send an event ID based on the current timestamp

“noEventId” - do not send an event ID to the IIS server.

pub fn set_event_id_mode(self, input: Option<SmoothGroupEventIdMode>) -> Self

pub fn event_stop_behavior(self, input: SmoothGroupEventStopBehavior) -> Self

When set to sendEos, send EOS signal to IIS server when stopping the event

pub fn set_event_stop_behavior(
self,
input: Option<SmoothGroupEventStopBehavior>
) -> Self

pub fn filecache_duration(self, input: i32) -> Self

Size in seconds of file cache for streaming outputs.

pub fn set_filecache_duration(self, input: Option<i32>) -> Self

pub fn fragment_length(self, input: i32) -> Self

Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

pub fn set_fragment_length(self, input: Option<i32>) -> Self

pub fn input_loss_action(self, input: InputLossActionForMsSmoothOut) -> Self

Parameter that control output group behavior on input loss.

pub fn set_input_loss_action(
self,
input: Option<InputLossActionForMsSmoothOut>
) -> Self

pub fn num_retries(self, input: i32) -> Self

Number of retry attempts.

pub fn set_num_retries(self, input: Option<i32>) -> Self

pub fn restart_delay(self, input: i32) -> Self

Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

pub fn set_restart_delay(self, input: Option<i32>) -> Self

pub fn segmentation_mode(self, input: SmoothGroupSegmentationMode) -> Self

useInputSegmentation has been deprecated. The configured segment size is always used.

pub fn set_segmentation_mode(
self,
input: Option<SmoothGroupSegmentationMode>
) -> Self

pub fn send_delay_ms(self, input: i32) -> Self

Number of milliseconds to delay the output from the second pipeline.

pub fn set_send_delay_ms(self, input: Option<i32>) -> Self

pub fn sparse_track_type(self, input: SmoothGroupSparseTrackType) -> Self

Identifies the type of data to place in the sparse track:

SCTE35: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame to start a new segment.

SCTE35_WITHOUT_SEGMENTATION: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame but don’t start a new segment.

NONE: Don’t generate a sparse track for any outputs in this output group.

pub fn set_sparse_track_type(
self,
input: Option<SmoothGroupSparseTrackType>
) -> Self

pub fn stream_manifest_behavior(
self,
input: SmoothGroupStreamManifestBehavior
) -> Self

When set to send, send stream manifest so publishing point doesn’t start until all streams start.

pub fn set_stream_manifest_behavior(
self,
input: Option<SmoothGroupStreamManifestBehavior>
) -> Self

pub fn timestamp_offset(self, input: impl Into<String>) -> Self

Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

pub fn set_timestamp_offset(self, input: Option<String>) -> Self

pub fn timestamp_offset_mode(
self,
input: SmoothGroupTimestampOffsetMode
) -> Self

Type of timestamp date offset to use.

useEventStartDate: Use the date the event was started as the offset

useConfiguredOffset: Use an explicitly configured date as the offset

pub fn set_timestamp_offset_mode(
self,
input: Option<SmoothGroupTimestampOffsetMode>
) -> Self

pub fn build(self) -> MsSmoothGroupSettings

Consumes the builder and constructs a MsSmoothGroupSettings