Struct aws_sdk_medialive::model::H265Settings

#[non_exhaustive]
pub struct H265Settings {

    pub adaptive_quantization: Option<H265AdaptiveQuantization>,
    pub afd_signaling: Option<AfdSignaling>,
    pub alternative_transfer_function: Option<H265AlternativeTransferFunction>,
    pub bitrate: i32,
    pub buf_size: i32,
    pub color_metadata: Option<H265ColorMetadata>,
    pub color_space_settings: Option<H265ColorSpaceSettings>,
    pub filter_settings: Option<H265FilterSettings>,
    pub fixed_afd: Option<FixedAfd>,
    pub flicker_aq: Option<H265FlickerAq>,
    pub framerate_denominator: i32,
    pub framerate_numerator: i32,
    pub gop_closed_cadence: i32,
    pub gop_size: f64,
    pub gop_size_units: Option<H265GopSizeUnits>,
    pub level: Option<H265Level>,
    pub look_ahead_rate_control: Option<H265LookAheadRateControl>,
    pub max_bitrate: i32,
    pub min_i_interval: i32,
    pub par_denominator: i32,
    pub par_numerator: i32,
    pub profile: Option<H265Profile>,
    pub qvbr_quality_level: i32,
    pub rate_control_mode: Option<H265RateControlMode>,
    pub scan_type: Option<H265ScanType>,
    pub scene_change_detect: Option<H265SceneChangeDetect>,
    pub slices: i32,
    pub tier: Option<H265Tier>,
    pub timecode_insertion: Option<H265TimecodeInsertionBehavior>,

}

H265 Settings

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

adaptive_quantization: Option<H265AdaptiveQuantization>

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

afd_signaling: Option<AfdSignaling>

Indicates that AFD values will be written into the output stream. If afdSignaling is “auto”, the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to “fixed”, the AFD value will be the value configured in the fixedAfd parameter.

alternative_transfer_function: Option<H265AlternativeTransferFunction>

Whether or not EML should insert an Alternative Transfer Function SEI message to support backwards compatibility with non-HDR decoders and displays.

bitrate: i32

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

buf_size: i32

Size of buffer (HRD buffer model) in bits.

color_metadata: Option<H265ColorMetadata>

Includes colorspace metadata in the output.

color_space_settings: Option<H265ColorSpaceSettings>

Color Space settings

filter_settings: Option<H265FilterSettings>

Optional filters that you can apply to an encode.

fixed_afd: Option<FixedAfd>

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to ‘Fixed’.

flicker_aq: Option<H265FlickerAq>

If set to enabled, adjust quantization within each frame to reduce flicker or ‘pop’ on I-frames.

framerate_denominator: i32

Framerate denominator.

framerate_numerator: i32

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

gop_closed_cadence: i32

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

gop_size: f64

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

gop_size_units: Option<H265GopSizeUnits>

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

level: Option<H265Level>

H.265 Level.

look_ahead_rate_control: Option<H265LookAheadRateControl>

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

max_bitrate: i32

For QVBR: See the tooltip for Quality level

min_i_interval: i32

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

par_denominator: i32

Pixel Aspect Ratio denominator.

par_numerator: i32

Pixel Aspect Ratio numerator.

profile: Option<H265Profile>

H.265 Profile.

qvbr_quality_level: i32

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. Set values for the QVBR quality level field and Max bitrate field that suit your most important viewing devices. Recommended values are:

Primary screen: Quality level: 8 to 10. Max bitrate: 4M

PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M

Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M

rate_control_mode: Option<H265RateControlMode>

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

scan_type: Option<H265ScanType>

Sets the scan type of the output to progressive or top-field-first interlaced.

scene_change_detect: Option<H265SceneChangeDetect>

Scene change detection.

slices: i32

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

tier: Option<H265Tier>

H.265 Tier.

timecode_insertion: Option<H265TimecodeInsertionBehavior>

Determines how timecodes should be inserted into the video elementary stream.

‘disabled’: Do not include timecodes

‘picTimingSei’: Pass through picture timing SEI messages from the source specified in Timecode Config

Implementations

impl H265Settings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture H265Settings