Struct aws_sdk_medialive::model::M2tsSettings

#[non_exhaustive]
pub struct M2tsSettings {

    pub absent_input_audio_behavior: Option<M2tsAbsentInputAudioBehavior>,
    pub arib: Option<M2tsArib>,
    pub arib_captions_pid: Option<String>,
    pub arib_captions_pid_control: Option<M2tsAribCaptionsPidControl>,
    pub audio_buffer_model: Option<M2tsAudioBufferModel>,
    pub audio_frames_per_pes: i32,
    pub audio_pids: Option<String>,
    pub audio_stream_type: Option<M2tsAudioStreamType>,
    pub bitrate: i32,
    pub buffer_model: Option<M2tsBufferModel>,
    pub cc_descriptor: Option<M2tsCcDescriptor>,
    pub dvb_nit_settings: Option<DvbNitSettings>,
    pub dvb_sdt_settings: Option<DvbSdtSettings>,
    pub dvb_sub_pids: Option<String>,
    pub dvb_tdt_settings: Option<DvbTdtSettings>,
    pub dvb_teletext_pid: Option<String>,
    pub ebif: Option<M2tsEbifControl>,
    pub ebp_audio_interval: Option<M2tsAudioInterval>,
    pub ebp_lookahead_ms: i32,
    pub ebp_placement: Option<M2tsEbpPlacement>,
    pub ecm_pid: Option<String>,
    pub es_rate_in_pes: Option<M2tsEsRateInPes>,
    pub etv_platform_pid: Option<String>,
    pub etv_signal_pid: Option<String>,
    pub fragment_time: f64,
    pub klv: Option<M2tsKlv>,
    pub klv_data_pids: Option<String>,
    pub nielsen_id3_behavior: Option<M2tsNielsenId3Behavior>,
    pub null_packet_bitrate: f64,
    pub pat_interval: i32,
    pub pcr_control: Option<M2tsPcrControl>,
    pub pcr_period: i32,
    pub pcr_pid: Option<String>,
    pub pmt_interval: i32,
    pub pmt_pid: Option<String>,
    pub program_num: i32,
    pub rate_mode: Option<M2tsRateMode>,
    pub scte27_pids: Option<String>,
    pub scte35_control: Option<M2tsScte35Control>,
    pub scte35_pid: Option<String>,
    pub segmentation_markers: Option<M2tsSegmentationMarkers>,
    pub segmentation_style: Option<M2tsSegmentationStyle>,
    pub segmentation_time: f64,
    pub timed_metadata_behavior: Option<M2tsTimedMetadataBehavior>,
    pub timed_metadata_pid: Option<String>,
    pub transport_stream_id: i32,
    pub video_pid: Option<String>,

}

M2ts Settings

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

absent_input_audio_behavior: Option<M2tsAbsentInputAudioBehavior>

When set to drop, output audio streams will be removed from the program if the selected input audio stream is removed from the input. This allows the output audio configuration to dynamically change based on input configuration. If this is set to encodeSilence, all output audio streams will output encoded silence when not connected to an active input stream.

arib: Option<M2tsArib>

When set to enabled, uses ARIB-compliant field muxing and removes video descriptor.

arib_captions_pid: Option<String>

Packet Identifier (PID) for ARIB Captions in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

arib_captions_pid_control: Option<M2tsAribCaptionsPidControl>

If set to auto, pid number used for ARIB Captions will be auto-selected from unused pids. If set to useConfigured, ARIB Captions will be on the configured pid number.

audio_buffer_model: Option<M2tsAudioBufferModel>

When set to dvb, uses DVB buffer model for Dolby Digital audio. When set to atsc, the ATSC model is used.

audio_frames_per_pes: i32

The number of audio frames to insert for each PES packet.

audio_pids: Option<String>

Packet Identifier (PID) of the elementary audio stream(s) in the transport stream. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

audio_stream_type: Option<M2tsAudioStreamType>

When set to atsc, uses stream type = 0x81 for AC3 and stream type = 0x87 for EAC3. When set to dvb, uses stream type = 0x06.

bitrate: i32

The output bitrate of the transport stream in bits per second. Setting to 0 lets the muxer automatically determine the appropriate bitrate.

buffer_model: Option<M2tsBufferModel>

Controls the timing accuracy for output network traffic. Leave as MULTIPLEX to ensure accurate network packet timing. Or set to NONE, which might result in lower latency but will result in more variability in output network packet timing. This variability might cause interruptions, jitter, or bursty behavior in your playback or receiving devices.

cc_descriptor: Option<M2tsCcDescriptor>

When set to enabled, generates captionServiceDescriptor in PMT.

dvb_nit_settings: Option<DvbNitSettings>

Inserts DVB Network Information Table (NIT) at the specified table repetition interval.

dvb_sdt_settings: Option<DvbSdtSettings>

Inserts DVB Service Description Table (SDT) at the specified table repetition interval.

dvb_sub_pids: Option<String>

Packet Identifier (PID) for input source DVB Subtitle data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

dvb_tdt_settings: Option<DvbTdtSettings>

Inserts DVB Time and Date Table (TDT) at the specified table repetition interval.

dvb_teletext_pid: Option<String>

Packet Identifier (PID) for input source DVB Teletext data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

ebif: Option<M2tsEbifControl>

If set to passthrough, passes any EBIF data from the input source to this output.

ebp_audio_interval: Option<M2tsAudioInterval>

When videoAndFixedIntervals is selected, audio EBP markers will be added to partitions 3 and 4. The interval between these additional markers will be fixed, and will be slightly shorter than the video EBP marker interval. Only available when EBP Cablelabs segmentation markers are selected. Partitions 1 and 2 will always follow the video interval.

ebp_lookahead_ms: i32

When set, enforces that Encoder Boundary Points do not come within the specified time interval of each other by looking ahead at input video. If another EBP is going to come in within the specified time interval, the current EBP is not emitted, and the segment is “stretched” to the next marker. The lookahead value does not add latency to the system. The Live Event must be configured elsewhere to create sufficient latency to make the lookahead accurate.

ebp_placement: Option<M2tsEbpPlacement>

Controls placement of EBP on Audio PIDs. If set to videoAndAudioPids, EBP markers will be placed on the video PID and all audio PIDs. If set to videoPid, EBP markers will be placed on only the video PID.

ecm_pid: Option<String>

This field is unused and deprecated.

es_rate_in_pes: Option<M2tsEsRateInPes>

Include or exclude the ES Rate field in the PES header.

etv_platform_pid: Option<String>

Packet Identifier (PID) for input source ETV Platform data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

etv_signal_pid: Option<String>

Packet Identifier (PID) for input source ETV Signal data to this output. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

fragment_time: f64

The length in seconds of each fragment. Only used with EBP markers.

klv: Option<M2tsKlv>

If set to passthrough, passes any KLV data from the input source to this output.

klv_data_pids: Option<String>

Packet Identifier (PID) for input source KLV data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

nielsen_id3_behavior: Option<M2tsNielsenId3Behavior>

If set to passthrough, Nielsen inaudible tones for media tracking will be detected in the input audio and an equivalent ID3 tag will be inserted in the output.

null_packet_bitrate: f64

Value in bits per second of extra null packets to insert into the transport stream. This can be used if a downstream encryption system requires periodic null packets.

pat_interval: i32

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

pcr_control: Option<M2tsPcrControl>

When set to pcrEveryPesPacket, a Program Clock Reference value is inserted for every Packetized Elementary Stream (PES) header. This parameter is effective only when the PCR PID is the same as the video or audio elementary stream.

pcr_period: i32

Maximum time in milliseconds between Program Clock Reference (PCRs) inserted into the transport stream.

pcr_pid: Option<String>

Packet Identifier (PID) of the Program Clock Reference (PCR) in the transport stream. When no value is given, the encoder will assign the same value as the Video PID. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

pmt_interval: i32

The number of milliseconds between instances of this table in the output transport stream. Valid values are 0, 10..1000.

pmt_pid: Option<String>

Packet Identifier (PID) for the Program Map Table (PMT) in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

program_num: i32

The value of the program number field in the Program Map Table.

rate_mode: Option<M2tsRateMode>

When vbr, does not insert null packets into transport stream to fill specified bitrate. The bitrate setting acts as the maximum bitrate when vbr is set.

scte27_pids: Option<String>

Packet Identifier (PID) for input source SCTE-27 data to this output. Multiple values are accepted, and can be entered in ranges and/or by comma separation. Can be entered as decimal or hexadecimal values. Each PID specified must be in the range of 32 (or 0x20)..8182 (or 0x1ff6).

scte35_control: Option<M2tsScte35Control>

Optionally pass SCTE-35 signals from the input source to this output.

scte35_pid: Option<String>

Packet Identifier (PID) of the SCTE-35 stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

segmentation_markers: Option<M2tsSegmentationMarkers>

Inserts segmentation markers at each segmentationTime period. raiSegstart sets the Random Access Indicator bit in the adaptation field. raiAdapt sets the RAI bit and adds the current timecode in the private data bytes. psiSegstart inserts PAT and PMT tables at the start of segments. ebp adds Encoder Boundary Point information to the adaptation field as per OpenCable specification OC-SP-EBP-I01-130118. ebpLegacy adds Encoder Boundary Point information to the adaptation field using a legacy proprietary format.

segmentation_style: Option<M2tsSegmentationStyle>

The segmentation style parameter controls how segmentation markers are inserted into the transport stream. With avails, it is possible that segments may be truncated, which can influence where future segmentation markers are inserted. When a segmentation style of “resetCadence” is selected and a segment is truncated due to an avail, we will reset the segmentation cadence. This means the subsequent segment will have a duration of $segmentationTime seconds. When a segmentation style of “maintainCadence” is selected and a segment is truncated due to an avail, we will not reset the segmentation cadence. This means the subsequent segment will likely be truncated as well. However, all segments after that will have a duration of $segmentationTime seconds. Note that EBP lookahead is a slight exception to this rule.

segmentation_time: f64

The length in seconds of each segment. Required unless markers is set to none.

timed_metadata_behavior: Option<M2tsTimedMetadataBehavior>

When set to passthrough, timed metadata will be passed through from input to output.

timed_metadata_pid: Option<String>

Packet Identifier (PID) of the timed metadata stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

transport_stream_id: i32

The value of the transport stream ID field in the Program Map Table.

video_pid: Option<String>

Packet Identifier (PID) of the elementary video stream in the transport stream. Can be entered as a decimal or hexadecimal value. Valid values are 32 (or 0x20)..8182 (or 0x1ff6).

Implementations

impl M2tsSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture M2tsSettings