Struct aws_sdk_medialive::model::MsSmoothGroupSettings

#[non_exhaustive]
pub struct MsSmoothGroupSettings {

    pub acquisition_point_id: Option<String>,
    pub audio_only_timecode_control: Option<SmoothGroupAudioOnlyTimecodeControl>,
    pub certificate_mode: Option<SmoothGroupCertificateMode>,
    pub connection_retry_interval: i32,
    pub destination: Option<OutputLocationRef>,
    pub event_id: Option<String>,
    pub event_id_mode: Option<SmoothGroupEventIdMode>,
    pub event_stop_behavior: Option<SmoothGroupEventStopBehavior>,
    pub filecache_duration: i32,
    pub fragment_length: i32,
    pub input_loss_action: Option<InputLossActionForMsSmoothOut>,
    pub num_retries: i32,
    pub restart_delay: i32,
    pub segmentation_mode: Option<SmoothGroupSegmentationMode>,
    pub send_delay_ms: i32,
    pub sparse_track_type: Option<SmoothGroupSparseTrackType>,
    pub stream_manifest_behavior: Option<SmoothGroupStreamManifestBehavior>,
    pub timestamp_offset: Option<String>,
    pub timestamp_offset_mode: Option<SmoothGroupTimestampOffsetMode>,

}

Ms Smooth Group Settings

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

acquisition_point_id: Option<String>

The ID to include in each message in the sparse track. Ignored if sparseTrackType is NONE.

audio_only_timecode_control: Option<SmoothGroupAudioOnlyTimecodeControl>

If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

certificate_mode: Option<SmoothGroupCertificateMode>

If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

connection_retry_interval: i32

Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

destination: Option<OutputLocationRef>

Smooth Streaming publish point on an IIS server. Elemental Live acts as a “Push” encoder to IIS.

event_id: Option<String>

MS Smooth event ID to be sent to the IIS server. Should only be specified if eventIdMode is set to useConfigured.

event_id_mode: Option<SmoothGroupEventIdMode>

Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run. Options:

“useConfigured” - use the value provided in eventId

“useTimestamp” - generate and send an event ID based on the current timestamp

“noEventId” - do not send an event ID to the IIS server.

event_stop_behavior: Option<SmoothGroupEventStopBehavior>

When set to sendEos, send EOS signal to IIS server when stopping the event

filecache_duration: i32

Size in seconds of file cache for streaming outputs.

fragment_length: i32

Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

input_loss_action: Option<InputLossActionForMsSmoothOut>

Parameter that control output group behavior on input loss.

num_retries: i32

Number of retry attempts.

restart_delay: i32

Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

segmentation_mode: Option<SmoothGroupSegmentationMode>

useInputSegmentation has been deprecated. The configured segment size is always used.

send_delay_ms: i32

Number of milliseconds to delay the output from the second pipeline.

sparse_track_type: Option<SmoothGroupSparseTrackType>

Identifies the type of data to place in the sparse track:

SCTE35: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame to start a new segment.

SCTE35_WITHOUT_SEGMENTATION: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame but don’t start a new segment.

NONE: Don’t generate a sparse track for any outputs in this output group.

stream_manifest_behavior: Option<SmoothGroupStreamManifestBehavior>

When set to send, send stream manifest so publishing point doesn’t start until all streams start.

timestamp_offset: Option<String>

Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

timestamp_offset_mode: Option<SmoothGroupTimestampOffsetMode>

Type of timestamp date offset to use.

useEventStartDate: Use the date the event was started as the offset

useConfiguredOffset: Use an explicitly configured date as the offset

Implementations

impl MsSmoothGroupSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture MsSmoothGroupSettings