Module aws_sdk_medialive::operation

Structs

AcceptInputDeviceTransfer

Accept an incoming input device transfer. The ownership of the device will transfer to your AWS account.

BatchDelete

Starts delete of resources.

BatchStart

Starts existing resources

BatchStop

Stops running resources

BatchUpdateSchedule

Update a channel schedule

CancelInputDeviceTransfer

Cancel an input device transfer that you have requested.

CreateChannel

Creates a new channel

CreateInput

Create an input

CreateInputSecurityGroup

Creates a Input Security Group

CreateMultiplex

Create a new multiplex.

CreateMultiplexProgram

Create a new program in the multiplex.

CreatePartnerInput

Create a partner input

CreateTags

Create tags for a resource

DeleteChannel

Starts deletion of channel. The associated outputs are also deleted.

DeleteInput

Deletes the input end point

DeleteInputSecurityGroup

Deletes an Input Security Group

DeleteMultiplex

Delete a multiplex. The multiplex must be idle.

DeleteMultiplexProgram

Delete a program from a multiplex.

DeleteReservation

Delete an expired reservation.

DeleteSchedule

Delete all schedule actions on a channel.

DeleteTags

Removes tags for a resource

DescribeChannel

Gets details about a channel

DescribeInput

Produces details about an input

DescribeInputDevice

Gets the details for the input device

DescribeInputDeviceThumbnail

Get the latest thumbnail data for the input device.

DescribeInputSecurityGroup

Produces a summary of an Input Security Group

DescribeMultiplex

Gets details about a multiplex.

DescribeMultiplexProgram

Get the details for a program in a multiplex.

DescribeOffering

Get details for an offering.

DescribeReservation

Get details for a reservation.

DescribeSchedule

Get a channel schedule

ListChannels

Produces list of channels that have been created

ListInputDeviceTransfers

List input devices that are currently being transferred. List input devices that you are transferring from your AWS account or input devices that another AWS account is transferring to you.

ListInputDevices

List input devices

ListInputSecurityGroups

Produces a list of Input Security Groups for an account

ListInputs

Produces list of inputs that have been created

ListMultiplexPrograms

List the programs that currently exist for a specific multiplex.

ListMultiplexes

Retrieve a list of the existing multiplexes.

ListOfferings

List offerings available for purchase.

ListReservations

List purchased reservations.

ListTagsForResource

Produces list of tags that have been created for a resource

PurchaseOffering

Purchase an offering and create a reservation.

RejectInputDeviceTransfer

Reject the transfer of the specified input device to your AWS account.

StartChannel

Starts an existing channel

StartMultiplex

Start (run) the multiplex. Starting the multiplex does not start the channels. You must explicitly start each channel.

StopChannel

Stops a running channel

StopMultiplex

Stops a running multiplex. If the multiplex isn’t running, this action has no effect.

TransferInputDevice

Start an input device transfer to another AWS account. After you make the request, the other account must accept or reject the transfer.

UpdateChannel

Updates a channel.

UpdateChannelClass

Changes the class of the channel.

UpdateInput

Updates an input.

UpdateInputDevice

Updates the parameters for the input device.

UpdateInputSecurityGroup

Update an Input Security Group’s Whilelists.

UpdateMultiplex

Updates a multiplex.

UpdateMultiplexProgram

Update a program in a multiplex.

UpdateReservation

Update reservation.